### 2022-04-10 #4309 @editorialbot
[PRE REVIEW]: Simmate: a framework for materials science
**Submitting author:** @jacksund (<a href="http://orcid.org/0000-0001-5739-8919">Jack Sundberg</a>)
**Repository:** https://github.com/jacksund/simmate
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.1
