### 2022-04-10 #4309 @jacksund
> Failed to discover a `Statement of need` section in paper

We modeled our paper after the [example one](https://joss.theoj.org/papers/10.21105/joss.00388) provided in the docs, where these sections were combined. If that is no longer allowed, we can change this though!
