### 2022-04-10 d7122ea
:bug: handle reset of non-simmate migrations
