+++ tutorials/08_Building_custom_workflows.md
@@ -1,0 +1,214 @@
+
+# Building custom workflows
+
+> :warning: This tutorial isn't required for beginners to use Simmate. There is currently no "quick tutorial" for this topic as this is only for advanced users with complex use-cases.
+
+In this tutorial, you will learn how to build customized workflows. This involves an introduction to underlying caculators and workflow engine.
+
+1. [Why isn't there a `custom_settings` option?](#why-isnt-there-a-custom_settings-option)
+2. [Create new & advanced workflows](#create-new-&-advanced-workflows)
+3. [Creating a Project for our workflow](#creating-a-project-for-our-workflow)
+
+</br>
+
+## Why isn't there a `custom_settings` option?
+
+> :bulb: we are testing out a "high customizable" workflow where you can change any settings via a `custom_settings` option in your input. Here, results will always be saved to the database, regardless of what you change. This is still in early testing though, so if you are interested in this feature, please reach out to our team.
+
+We intentionally avoid the use of `workflow.run(custom_settings=...)`. **This will NOT work.** Simmate does this because we do not want to store results from customized settings in the same results table -- as storing results from customized runs would (a) complicate analysis of many structuers/systems and (b) make navigating results extremely difficult for beginners. For example, reducing the `ENCUT` of a VASP calculation makes it so energies cannot be compared between all materials in the table, and thus, features like calculated hull energies would become inaccruate. Instead, Simmate encourages the creation of new workflows and result tables when you want to customize settings. This puts Simmate's emphasis on "scaling up" workflows (i.e. running a fixed workflow on thousands on materials) as opposed to "scaling out" workflows (i.e. a flexible workflow that changes on a structure-by-structure basis). 
+
+If you wish to customize settings for a few calculations, we recommend using this approach:
+
+``` bash
+# This simply writes input files
+simmate workflows setup-only static-energy/mit POSCAR
+
+# access your files in the new directory
+cd MIT_Static_Energy_inputs
+
+# Customize input files as you see fit.
+# For example, you may want to edit INCAR settings
+nano INCAR
+
+# You can then submit VASP manually. Note, this will not use
+# simmate at all! So there is no error handling and no results
+# will be saved to your database.
+vasp_std > vasp.out
+```
+
+If you are submitting many calculations (>20) and this process doesn't suit your needs, keep reading!
+
+</br>
+
+## Create new & advanced workflows
+
+1. Simmate defines a base `Workflow` class to help with common material science analyses, but at its core, it works exactly the same as Prefect workflows. Therefore, you can take advantage of all of [Prefect's guides](https://docs.prefect.io/core/) when building new workflows. The simplest possible workflow can look something like...
+
+``` python
+from simmate.workflow_engine import Workflow, task
+
+@task
+def hello_task():
+    print("Hello World!")
+
+with Workflow("hello-flow") as my_new_workflow:
+    hello_task()
+
+my_new_workflow.run()
+```
+
+2.  In building a material science workflow, we may want two common steps: (a) running a calculation through a program like VASP and (b) saving the results to our database. We must therefore define individual `Task`s as well as a `DatabaseTable` to store results in. The next steps will address these.
+
+3. To build a simple task, we use the `@task` decorator like we did above with our `hello_task` function. However, because material science frequently involves calling external programs, we've created a powerful `S3Task` class to inherit from. A `S3Task` class and implement their own `setup`, `execute`, and `workup` methods, and even add functions for error handling. Here's how you would build a custom `S3Task` from scratch:
+
+``` python
+import os
+
[…]
