+++ tutorials/09_Add_custom_workflows.md
@@ -1,97 +1,0 @@
-:warning::warning::warning: THIS TUTORIAL IS INCOMPLETE AT THE MOMENT :warning::warning::warning:
-
-# Introduction to caculators and the workflow engine
-
-> :warning: This tutorial isn't required for beginners to use Simmate. Instead, this helps you understand what Simmate is doing behind the scenes. If you'd like to create custom workflows or contribute to Simmate, then understanding these core concepts is the place to start.
-
-# The quick tutorial
-
-1. If you'd like to see what third-party programs Simmate supports, take a look at the `simmate.calculators` module ([here](https://github.com/jacksund/simmate/tree/main/src/simmate/calculators))
-
-2. Each call to a program is considered a single task. These tasks all inherit from a `S3Task` class and implement their own `setup`, `execute`, and `workup` methods. Here's what top-level usage looks like for [VASP](https://vasp.at/):
-
-```python
-from simmate.calculators.vasp.tasks.base import VaspTask
-from simmate.calculators.vasp.inputs.potcar_mappings import PBE_ELEMENT_MAPPINGS_LOW_QUALITY
-
-class ExampleRelaxation(VaspTask):
-    functional = "PBE"
-    potcar_mappings = PBE_ELEMENT_MAPPINGS_LOW_QUALITY
-    incar = dict(
-        PREC="Low",
-        EDIFF=2e-3,
-        EDIFFG=-2e-1,
-        NSW=75,
-        IBRION=2,
-        POTIM=0.02,
-        KSPACING=0.75,
-    )
- 
-my_task = ExampleRelaxation()
-result = my_task.run()  # calls the setup, execute, and workup methods for us
-```
-
-3. Workflows add orchestration and database saving to the `S3Task`. The simplest workflows are just one call to a calculator:
-
-```python
-from simmate.workflow_engine.utilities import s3task_to_workflow
-from simmate.calculators.vasp.tasks.relaxation.example import ExampleRelaxationTask
-from simmate.calculators.vasp.database.example import ExampleRelaxationResults
-
-workflow = s3task_to_workflow(
-    name="Quality 00 Relaxation",
-    module=__name__,
-    project_name="Simmate-Relaxation",
-    s3task=ExampleRelaxationTask,
-    calculation_table=Quality00RelaxationResults,
-    register_kwargs=["prefect_flow_run_id", "structure", "source"],
-)
-
-result = workflow.run(structure=structure)  # now includes saving results to a database table!
-```
-
-*note: we keep `S3Task`'s and `Workflows` separate to for modularity of our code*
-
-4. Make a custom `S3task` whenever you want to test out new calculations. Results won't be saved to your database. When you're ready to start storing data from your tasks, ... TODO need to combine this tutorial with custom apps
-
-
-
-
-# The full tutorial
-
-## What is a `calculator`?
-
-A `Calculator` is an external program that Simmate calls in order to run a calculation (e.g. [VASP](https://vasp.at/), [ABINIT](https://www.abinit.org/), [bader](http://theory.cm.utexas.edu/henkelman/code/bader/), etc.) and the `simmate.calculators` module stores all of the code for input/outputs, common error handlers, and more.
-
-## What is an `S3Task`?
-
-A single calculator run is carried out by a `StagedSupervisedShellTask` or a "`S3Task`". There is some history behind why it's named this way, but here is how the name breaks down:
-
-- `Staged`: the overall calculation is made of three stages (each is a class method)
-    1. `setup` = write input files
-    2. `execution` = the actual running & monitoring of the program  <--- `Supervised` and `Shell` help describe this stage
-    3. `workup` = reading the output files
-- `Shell`: the calculator is called through the command-line
[…]

[…]
