### 2022-04-10 c15b0cd
:tropical_fish: prep v0.1.3 release and run pdoc
