### 2022-04-11 16fb598
:wrench: update view tests
