+++ src/simmate/website/core/test/test_core_views.py
@@ -10,7 +10,7 @@ def test_home_view(client):
     # test initial view
     response = client.get("/")
     assert response.status_code == 200
-    assertTemplateUsed("home/home.html")
+    assertTemplateUsed(response, "home/home.html")
 
     # test submission of the form
     response = client.post(
@@ -27,16 +27,16 @@ def test_home_view(client):
 def test_extras_view(client):
     response = client.get("/extras/")
     assert response.status_code == 200
-    assertTemplateUsed("core_components/extras.html")
+    assertTemplateUsed(response, "core_components/extras.html")
 
 
 def test_contact_view(client):
     response = client.get("/contact/")
     assert response.status_code == 200
-    assertTemplateUsed("core_components/contact.html")
+    assertTemplateUsed(response, "core_components/contact.html")
 
 
 def test_about_view(client):
     response = client.get("/about/")
     assert response.status_code == 200
-    assertTemplateUsed("core_components/about.html")
+    assertTemplateUsed(response, "core_components/about.html")
