+++ src/simmate/website/workflows/test/test_workflows_views.py
@@ -3,62 +3,78 @@
 import pytest
 from pytest_django.asserts import assertTemplateUsed
 from django.urls import reverse
+
 from simmate.workflow_engine import Workflow
+from simmate.workflows.utilities import (
+    WORKFLOW_TYPES,
+    get_list_of_all_workflows,
+    get_workflow,
+)
+
+ALL_WORKFLOWS = get_list_of_all_workflows()
 
 
 def test_workflows_view(client):
     response = client.get("/workflows/")
     assert response.status_code == 200
-    assertTemplateUsed("workflows/all.html")
+    assertTemplateUsed(response, "workflows/all.html")
 
 
-def test_workflows_by_type_view(client):
-    response = client.get("/workflows/static-energy/")
+@pytest.mark.parametrize("workflow_type", WORKFLOW_TYPES)
+def test_workflows_by_type_view(client, workflow_type):
+
+    # grabs f"/workflows/{workflow_type}/"
+    url = reverse(
+        "workflows_by_type",
+        kwargs={"workflow_type": workflow_type},
+    )
+
+    response = client.get(url)
     assert response.status_code == 200
-    assertTemplateUsed("workflows/by_type.html")
-
-    response = client.get("/workflows/relaxation/")
-    assert response.status_code == 200
-    assertTemplateUsed("workflows/by_type.html")
+    assertTemplateUsed(response, "workflows/by_type.html")
 
 
 @pytest.mark.django_db
-def test_workflow_detail_view(client):
+@pytest.mark.parametrize("workflow_name", ALL_WORKFLOWS)
+def test_workflow_detail_view(client, workflow_name):
+
+    workflow = get_workflow(workflow_name)
 
     # list view
-    response = client.get("/workflows/static-energy/mit/")
+    # grabs f"/workflows/{type}/{name}/")
+    url = reverse(
+        "workflow_detail",
+        kwargs={
+            "workflow_type": workflow.type,
+            "workflow_name": workflow.name_short,
+        },
+    )
+
+    response = client.get(url)
     assert response.status_code == 200
-    assertTemplateUsed("workflows/detail.html")
+    assertTemplateUsed(response, "workflows/detail.html")
 
     # detail view - found
     # grabs... "/workflows/static-energy/mit/1/"
-    # url = reverse(
-    #     "workflow_run_detail",
-    #     kwargs={
-    #         "workflow_type": "static-energy",
-    #         "workflow_name": "mit",
-    #         "workflow_run_id": 1,
-    #     },
-    # )
-    # response = client.get(url)
-    # assert response.status_code == 200
-    # assertTemplateUsed("workflows/detail_run.html")
+    # TODO: how should I populate with test data?
 
     # detail view - not found
     # grabs... "/workflows/static-energy/mit/999/"
     url = reverse(
         "workflow_run_detail",
         kwargs={
-            "workflow_type": "static-energy",
-            "workflow_name": "mit",
+            "workflow_type": workflow.type,
+            "workflow_name": workflow.name_short,
             "workflow_run_id": 999,
         },
     )
+
     response = client.get(url)
     assert response.status_code == 404
-    assertTemplateUsed("workflows/detail_run.html")
+    # assertTemplateUsed(response, "workflows/detail_run.html")
 
 
+# @pytest.mark.parametrize("workflow_name", ALL_WORKFLOWS) # TODO
 @pytest.mark.django_db
 def test_workflow_submit_view(client, sample_structures, mocker):
 
@@ -75,7 +91,7 @@ def test_workflow_submit_view(client, sample_structures, mocker):
     # loading blank page
     response = client.get(url)
     assert response.status_code == 200
-    assertTemplateUsed("workflows/submit.html")
[…]
