### 2022-04-14 b90e13b
:fish: fix minor doc typos
