+++ src/simmate/website/core_components/base_api_view.py
@@ -1,0 +1,244 @@
+# -*- coding: utf-8 -*-
+
+from django.http import HttpRequest
+
+from rest_framework.viewsets import GenericViewSet
+
+# from rest_framework.generics import GenericAPIView
+from rest_framework.response import Response
+from rest_framework.serializers import ModelSerializer
+
+from simmate.website.core_components.filters import DatabaseTableFilter
+from simmate.database.base_data_types import DatabaseTable
+
+
+# Consider inheriting from GenericViewSet in the future:
+# https://www.django-rest-framework.org/api-guide/viewsets/
+# https://github.com/encode/django-rest-framework/blob/master/rest_framework/viewsets.py
+
+
+class SimmateAPIViewSet(GenericViewSet):
+    """
+    Example use:
+
+    ``` python
+    # in views.py
+    class ExampleTableViewSet(SimmateAPIViewSet):
+        table = ExampleTable
+        template_list = "custom_app/list.html"
+        template_retrieve = "custom_app/retrieve.html"
+
+    # in urls.py
+    path(
+        route="example-table",
+        view=ExampleTableViewSet.list_view,
+        name="example-list",
+    )
+    path(
+        route="example-table/<int:id>",
+        view=ExampleTableViewSet.retrieve_view,
+        name="example-retrieve",
+    )
+    ```
+    """
+
+    table: DatabaseTable = None
+
+    template_list: str = None
+
+    template_retrieve: str = None
+
+    def get_list_response(self, request: HttpRequest, *args, **kwargs) -> Response:
+
+        # This code is modified from the ListModelMixin, where instead of returning
+        # a response, we perform additional introspection first.
+        # https://github.com/encode/django-rest-framework/blob/master/rest_framework/mixins.py#L33
+
+        # self.format_kwarg --> not sure why this always returns None, so I
+        # grab the format from the request instead. If it isn't listed, then
+        # I'm using the default which is html.
+        self._format_kwarg = request.GET.get("format", "html")
+
+        queryset = self.filter_queryset(self.get_queryset())
+
+        # attempt to paginate the results
+        page = self.paginate_queryset(queryset)
+        if page is not None:
+            serializer = self.get_serializer(page, many=True)
+        else:
+            serializer = self.get_serializer(queryset, many=True)
+
+        # If don't have the html format, we follow simple logic from the
+        # original ListModelMixin method
+        if self._format_kwarg != "html":
+            if page is not None:
+                return self.get_paginated_response(serializer.data)
+            else:
+                return Response(serializer.data)
+
+        # otherwise we assume the html format.
+        else:
+            filterset = self.filterset_class(request.GET)
+            data = {
+                # "filterset": filterset, # not used at the momemnt
+                "filterset_mixins": filterset.get_mixin_names(),
+                "form": filterset.form,
+                "extra_filters": filterset.get_extra_filters(),
+                "calculations": serializer.instance,  # return python objs, not dict
+                "ncalculations_matching": queryset.count(),
+                "ncalculations_possible": self.get_queryset().count(),
+                **self.paginator.get_html_context(),
+                **self.get_list_context(request, **kwargs),
+            }
+            return Response(data)
+
+    def get_retrieve_response(self, request: HttpRequest, *args, **kwargs) -> Response:
+
+        # self.format_kwarg --> not sure why this always returns None, so I
+        # grab the format from the request instead. If it isn't listed, then
+        # I'm using the default which is html.
+        self._format_kwarg = request.GET.get("format", "html")
+
+        # ---------------------------------------------------
+        # This code is from the RetrieveModelMixin, where instead of returning
+        # a response, we perform additional introspection first
+        instance = self.get_object()
+        serializer = self.get_serializer(instance)
+        # return Response(serializer.data) <--- removed from original
[…]
