+++ CHANGELOG.md
@@ -1,1 +1,24 @@
-# TODO
+# Release History
+
+Simmate releases follow [semantic versioning](https://semver.org/). This means versions numbers (e.g. v1.2.3) correspond to MAJOR.MINOR.PATCH:
+
+  - MAJOR version when you make incompatible API changes
+  - MINOR version when you add functionality in a backwards compatible manner
+  - PATCH version when you make backwards compatible bug fixes
+
+There is one key exception to the rules above -- and that is with MAJOR=0 releases. Any v0.x.y release is considered developmental where APIs are subject to change and should not be considered stable. *(This is consistent with the semantic version spec. See [point 4](https://semver.org/#spec-item-4).)*
+
+</br></br>
+
+# Unreleased
+- start the CHANGELOG!
+- refactor API views and add `SimmateAPIViewSet` class
+- refactor `simmate start-project` command and underlying methods
+- refactor `simmate workflow-engine run-cluster` command and underlying methods
+
+
+# Simmate v0.0.0 - v0.1.4
+- initial release
+- web interface styling
+- minor bug fixes
+- adding tests and docs
