+++ src/simmate/toolkit/base_data_types/structure.py
@@ -105,109 +105,24 @@ class Structure(PymatgenStructure):
 
         return structure_cleaned
 
+    # -------------------------------------------------------------------------
+
+    # All methods below wrap converters from the `simmate.file_converters.structure`
+    # module. To keep this class modular (from the database and optional deps),
+    # imports are done lazily. This means we import the relevant convert once
+    # the method is called -- rather than when this module is initially loaded.
+
     @classmethod
     def from_database(cls, structure: dict):
-        """
-        This is an experimental feature.
+        from simmate.file_converters.structure.database import DatabaseAdapter
 
-        Loads a structure from the Simmate database.
-        """
-
-        # because the structure is in the database, we need to setup django and
-        # make sure we can access the tables. This import break the modularity
-        # of this toolkit, so we therefore keep import within this function.
-        from simmate.configuration.django import setup_full  # connects to database
-        from simmate.website.workflows import models as all_datatables
-        from django.utils.module_loading import import_string
-
-        # start by loading the datbase table, which is given as a module path
-        datatable_str = structure["calculation_table"]
-
-        # Import the datatable class -- how this is done depends on if it
-        # is from a simmate supplied class or if the user supplied a full
-        # path to the class
-        # OPTIMIZE: is there a better way to do this?
-        if hasattr(all_datatables, datatable_str):
-            datatable = getattr(all_datatables, datatable_str)
-        else:
-            datatable = import_string(datatable_str)
-
-        # These attributes tells us which structure to grab from our datatable.
-        # The user should have only provided one -- if they gave more, we just
-        # use whichever one comes first.
-        prefect_flow_run_id = structure.get("prefect_flow_run_id")
-        calculation_id = structure.get("calculation_id")
-        directory_old = structure.get("directory")
-
-        # we must have either a prefect_flow_run_id or calculation_id
-        if not prefect_flow_run_id and not calculation_id and not directory_old:
-            raise Exception(
-                "You must have either a prefect_flow_run_id, calculation_id, "
-                "or directory provided if you want to load a structure from "
-                "a previous calculation."
-            )
-
-        # now query the datable with which whichever was provided. Each of these
-        # are unique so all three should return a single calculation.
-        if calculation_id:
-            calculation = datatable.objects.get(id=calculation_id)
-        elif prefect_flow_run_id:
-            calculation = datatable.objects.get(
-                prefect_flow_run_id=prefect_flow_run_id,
-            )
-        elif directory_old:
-            calculation = datatable.objects.get(directory=directory_old)
-
-        # In some cases, the structure we want is not within the calculation table.
-        # For example, in relaxations the final structure is attached via
-        # the table.structure_final attribute
-        structure_field = structure.get("structure_field")
-        if structure_field:
-            structure_cleaned = getattr(calculation, structure_field).to_toolkit()
-        # if there's no structure field, that means we already have the correct entry
-        else:
-            structure_cleaned = calculation.to_toolkit()
-
-        # For ease of access, we also link the calculation database entry
-        structure_cleaned.calculation = calculation
-
-        return structure_cleaned
+        return DatabaseAdapter.get_toolkit_from_database_dict(structure)
 
     @classmethod
     def from_database_string(cls, structure_string: str):
-        """
-        Loads a toolkit structure from a string -- specifically strings that
-        are stored in the structure_string column for
-        simmate.database.base_data_types.Structure.
-        """
-        # Dev note: this method should be merged with the Toolkit.from_str method.
-        # I only have this separate for now because pymatgen's from_str doesn't
-        # dynamically determine format from the string alone.
[…]
