+++ src/simmate/file_converters/voxeldata/README.md
@@ -1,0 +1,5 @@
+
+VoxelData Converters
+--------------------
+
+Simmate still focuses almost exclusively on bulk crystal structures, so this module is empty at the moment. However, we keep this here for outlining and planning purposes. In the future, this will be a library of converters for volume-based data files and object types.
