+++ src/simmate/file_converters/README.md
@@ -2,6 +2,6 @@
 File Converters
 ---------------
 
-This module hosts functionality for reading and writing different file type. For example, there are many ways to represent a crystal structure -- such as the POSCAR, CASTEP, and CIF file formats -- and it therefore important to have functions ready to convert between the two. Beyond file types, this module also include functions for converting between object types (i.e. different python classes).
+This module hosts functionality for reading and writing different file types. For example, there are many ways to represent a crystal structure -- such as the POSCAR, CASTEP, and CIF file formats -- and it's therefore important to have functions to convert between the them. Beyond file types, this module also include functions for converting between object types (i.e. different python classes).
 
 NOTE: Currently, this module only hosts converters for `structure`, whereas the `molecule` and `voxeldata` types are still under planning/development.
