### 2022-04-17 98c42f4
:olive: standardize energy to static_energy
