+++ src/simmate/calculators/vasp/workflows/static_energy/materials_project.py
@@ -1,18 +1,18 @@
 # -*- coding: utf-8 -*-
 
 from simmate.workflow_engine import s3task_to_workflow
-from simmate.calculators.vasp.tasks.energy import (
-    Quality04Energy as Quality04EnergyTask,
+from simmate.calculators.vasp.tasks.static_energy import (
+    MatProjStaticEnergy as MPStaticEnergyTask,
 )
 from simmate.calculators.vasp.database.energy import (
-    Quality04StaticEnergy as Quality04StaticEnergyResults,
+    MatProjStaticEnergy as MPStaticEnergyResults,
 )
 
 workflow = s3task_to_workflow(
-    name="static-energy/quality04",
+    name="static-energy/matproj",
     module=__name__,
     project_name="Simmate-Energy",
-    s3task=Quality04EnergyTask,
-    calculation_table=Quality04StaticEnergyResults,
+    s3task=MPStaticEnergyTask,
+    calculation_table=MPStaticEnergyResults,
     register_kwargs=["prefect_flow_run_id", "structure", "source"],
 )
