+++ src/simmate/calculators/vasp/tasks/density_of_states.py
@@ -4,7 +4,7 @@ import os
 
 from pymatgen.electronic_structure.plotter import DosPlotter
 
-from simmate.calculators.vasp.tasks.energy.materials_project import (
+from simmate.calculators.vasp.tasks.static_energy.materials_project import (
     MatProjStaticEnergy,
 )
 
