+++ src/simmate/calculators/vasp/tasks/population_analysis/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .pre_bader import MatProjPreBader
