+++ src/simmate/calculators/vasp/workflows/population_analysis/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .bader import workflow as bader_matproj_workflow
