+++ src/simmate/calculators/vasp/tasks/population_analysis/__init__.py
@@ -1,3 +1,5 @@
 # -*- coding: utf-8 -*-
 
 from .pre_bader import MatProjPreBader
+from .elf import MatProjELF
+from .elf_bader import MatProjPreBaderELF
