+++ src/simmate/workflows/population_analysis.py
@@ -2,5 +2,6 @@
 
 # TODO: when I add more calculators, I can do something like this...
 # if "simmate.calculators.vasp" in installed_apps:
-
-# TODO
+from simmate.calculators.vasp.workflows.population_analysis import (
+    bader_matproj_workflow,
+)
