+++ src/simmate/workflows/test/test_workflows_utilities.py
@@ -30,6 +30,7 @@ def test_list_of_all_workflows():
         "relaxation/quality03",
         "relaxation/quality04",
         "relaxation/staged",
+        "population-analysis/bader-matproj",
         "band-structure/matproj",
         "density-of-states/matproj",
         "dynamics/mit",
