+++ src/simmate/workflows/population_analysis.py
@@ -4,4 +4,5 @@
 # if "simmate.calculators.vasp" in installed_apps:
 from simmate.calculators.vasp.workflows.population_analysis import (
     bader_matproj_workflow,
+    elf_matproj_workflow,
 )
