+++ src/simmate/workflows/test/test_workflows_utilities.py
@@ -31,6 +31,7 @@ def test_list_of_all_workflows():
         "relaxation/quality04",
         "relaxation/staged",
         "population-analysis/bader-matproj",
+        "population-analysis/elf-matproj",
         "band-structure/matproj",
         "density-of-states/matproj",
         "dynamics/mit",
