+++ src/simmate/calculators/vasp/workflows/population_analysis/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from .bader import workflow as bader_matproj_workflow
+from .elf import workflow as elf_matproj_workflow
