+++ src/simmate/workflows/customized.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from simmate.calculators.vasp.workflows.customized import vasp_workflow
