### 2022-04-19 d6af347
:rabbit: prep release and run pdoc
