### 2022-04-20 6a97a8a
:sparkles: archive when error handlers triggered
