+++ src/simmate/configuration/django/database.py
@@ -5,7 +5,7 @@ import shutil
 
 from django.core.management import call_command
 
-from simmate.configuration.django import setup_full  # sets database connection
+from simmate.database import connect
 from simmate.configuration.django.settings import DATABASES
 from simmate.database.base_data_types import Spacegroup
 
