### 2022-04-21 37eab9e
:sparkler: depreciate use_prev_dir in favor of copy_prev_dir
