### 2022-04-21 5820934
:prince: add django-allauth
