+++ tutorials/Guides_for_contributors/environment.yaml
@@ -16,6 +16,7 @@ dependencies:
     - dask-jobqueue >=0.7.3
     - dj-database-url >=0.5.0
     - django >=4.0.0
+    - django-allauth >=0.50.0
     - django-crispy-forms >=1.13.0
     - django-pandas >=0.6.6
     - django-filter >=21.1
