### 2022-04-22 f2b9ae7
:melon: fix typos in readme
