+++ src/simmate/calculators/vasp/workflows/static_energy/mit.py
@@ -15,4 +15,5 @@ workflow = s3task_to_workflow(
     s3task=MITStaticEnergyTask,
     calculation_table=MITStaticEnergyResults,
     register_kwargs=["prefect_flow_run_id", "structure", "source"],
+    description_doc_short="uses MIT Project settings",
 )
