+++ src/simmate/calculators/vasp/workflows/relaxation/materials_project.py
@@ -15,4 +15,5 @@ workflow = s3task_to_workflow(
     s3task=MPRelaxationTask,
     calculation_table=MPRelaxationResults,
     register_kwargs=["prefect_flow_run_id", "structure", "source"],
+    description_doc_short="uses Materials Project settings",
 )
