+++ src/simmate/workflow_engine/test/test_utilities.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# !!! The s3task_to_workflow function requires a Django database table, so
-# I need to figure out pytest-django before writing a test for it.
