+++ src/simmate/calculators/vasp/workflows/static_energy/neb_endpoint.py
@@ -15,4 +15,5 @@ workflow = s3task_to_workflow(
     s3task=NEBEndpointStaticEnergyTask,
     calculation_table=NEBEndpointStaticEnergyResults,
     register_kwargs=["prefect_flow_run_id", "structure", "source"],
+    description_doc_short="uses Materials Project settings meant for defect supercell structures",
 )
