### 2022-04-23 46313c0
:mag: add show_parameters features
