+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -2,13 +2,14 @@
 name: Bug report
 about: Create a report to help us improve
 title: ''
-labels: ''
+labels: 'bug'
 assignees: ''
 
 ---
 
 ## Describe the bug
 A clear and concise description of what the bug is.
+
 
 ## To Reproduce
 Steps to reproduce the behavior:
@@ -34,8 +35,10 @@ structure = Structure.from_file(...)
 < keep this as short and clear as possible! >
 ```
 
+
 ## Expected behavior
 A clear and concise description of what you expected to happen.
+
 
 ## Additional context
 Add any other context about the problem here, such as relevant links that make help implement a solution. (optional)
