### 2022-04-25 73e6be3
:eyes: update lmaxmix to mp settings
