### 2022-04-27 e3e5e04
:otter: add pymatgen-set compatibility tests
