+++ src/simmate/calculators/vasp/tasks/relaxation/materials_project_hse.py
@@ -41,40 +41,37 @@ from simmate.calculators.vasp.error_handlers import (
 )
 
 
-class MatProjRelaxation(VaspTask):
+class MatProjHSERelaxation(VaspTask):
     """
+    Runs a VASP relaxation calculation using Materials Project HSE settings.
+
     This task is a reimplementation of pymatgen's
-    [MPRelaxSet](https://pymatgen.org/pymatgen.io.vasp.sets.html#pymatgen.io.vasp.sets.MPRelaxSet).
-
-    Runs a VASP geometry optimization using Materials Project settings.
-
-    Materials Project settings are often considered the minimum-required
-    quality for publication and is sufficient for most applications. If you are
-    looking at one structure in detail (for electronic, vibrational, and other
-    properties), you should still test for convergence using higher-quality
-    settings.
+    [MPHSERelaxSet](https://pymatgen.org/pymatgen.io.vasp.sets.html#pymatgen.io.vasp.sets.MPHSERelaxSet).
     """
 
     functional = "PBE"
     potcar_mappings = PBE_ELEMENT_MAPPINGS
 
     incar = dict(
-        ALGO="Fast",
+        ALGO="All",
         EDIFF__per_atom=5.0e-05,
         ENCUT=520,
+        HFSCREEN=0.2,
         IBRION=2,
+        ICHARG=1,
         ISIF=3,
-        ISMEAR=-5,
+        ISMEAR=0,
         ISPIN=2,
-        KSPACING=0.4,  # !!! This is where we are different from pymatgen right now
-        LASPH=True,
+        LHFCALC=True,
         LORBIT=11,
-        LREAL="Auto",
+        LREAL="AUTO",
         LWAVE=False,
         NELM=100,
         NSW=99,
         PREC="Accurate",
+        PRECFOCK="Fast",
         SIGMA=0.05,
+        KSPACING=0.4,  # !!! This is where we are different from pymatgen right now
         MAGMOM__smart_magmom={
             "default": 0.6,  # note this is different from the VASP default
             "Ce": 5,
@@ -109,56 +106,6 @@ class MatProjRelaxation(VaspTask):
             "Yb3+": 1,
         },
         LMAXMIX__smart_lmaxmix=True,
-        multiple_keywords__smart_ldau=dict(
-            LDAU__auto=True,
-            LDAUTYPE=2,
-            LDAUPRINT=1,
-            LDAUJ={},  # pymatgen sets these, but they're all default values anyways
-            LDAUL={
-                "F": {
-                    "Co": 2,
-                    "Cr": 2,
-                    "Fe": 2,
-                    "Mn": 2,
-                    "Mo": 2,
-                    "Ni": 2,
-                    "V": 2,
-                    "W": 2,
-                },
-                "O": {
-                    "Co": 2,
-                    "Cr": 2,
-                    "Fe": 2,
-                    "Mn": 2,
-                    "Mo": 2,
-                    "Ni": 2,
-                    "V": 2,
-                    "W": 2,
-                },
-            },
-            LDAUU={
-                "F": {
-                    "Co": 3.32,
-                    "Cr": 3.7,
-                    "Fe": 5.3,
-                    "Mn": 3.9,
-                    "Mo": 4.38,
-                    "Ni": 6.2,
-                    "V": 3.25,
-                    "W": 6.2,
-                },
-                "O": {
-                    "Co": 3.32,
-                    "Cr": 3.7,
-                    "Fe": 5.3,
-                    "Mn": 3.9,
-                    "Mo": 4.38,
-                    "Ni": 6.2,
-                    "V": 3.25,
-                    "W": 6.2,
-                },
-            },
-        ),
     )
 
     error_handlers = [
