+++ src/simmate/calculators/vasp/tasks/nudged_elastic_band/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from .base import VaspNudgedElasticBandTask
+
+from .materials_virtual_lab_ci import MatVirtualLabCINEB
+from .mit import MITNudgedElasticBand
