+++ src/simmate/calculators/vasp/tasks/relaxation/__init__.py
@@ -7,4 +7,6 @@ from .quality_03 import Quality03Relaxation
 from .quality_04 import Quality04Relaxation
 from .mit import MITRelaxation
 from .materials_project import MatProjRelaxation
-from .neb_endpoint import NEBEndpointRelaxation
+from .materials_virtual_lab_endpoint import (
+    MatVirtualLabCINEBEndpointRelaxation as NEBEndpointRelaxation,
+)
