### 2022-04-27 c4893c8
:fox_face: continue converting pymatgen sets
