+++ src/simmate/calculators/vasp/tasks/band_structure/base.py
@@ -12,40 +12,26 @@ from simmate.calculators.vasp.tasks.static_energy.materials_project import (
 )
 
 
-class MatProjBandStructure(MatProjStaticEnergy):
+class VaspBandStructure(MatProjStaticEnergy):
     """
-    Calculates the band structure using Materials Project settings.
+    A base class for band structure calculations. This is not meant
+    to be used directly but instead should be inherited from.
 
     Your stucture will be converted to the standardized-primitive unitcell so
     that the high-symmetry K-point path can be used.
 
     This is also a non self-consistent field (non-SCF) calculation and thus uses
     the a fixed charge density from a previous static energy calculation.
-
-    We do NOT recommend running this calculation on its own. Instead, you should
-    use the full workflow, which handles the preliminary relaxation and SCF
-    energy calculation for you. This S3Task is only the final step of that workflow.
-
-    See `vasp.workflows.band_structure`.
     """
-
-    # The default settings to use for this calculation.
-    # The key thing for band structures is that this follows a static energy
-    # calculation. We use the CHGCAR from that previous calculation. We also
-    # hold the charge density static during this calculation so this is a
-    # non-selfconsistent (non-SCF) calculation.
-    incar = MatProjStaticEnergy.incar.copy()
-    incar.update(
-        ICHARG=11,
-        ISMEAR=0,
-        SIGMA=0.01,
-    )
 
     # set the KptGrid or KptPath object
     # TODO: in the future, all functionality of this class will be available
     # by giving a KptPath class here.
     kpoints = None
     kpoints_line_density = 20
+    """
+    Density of k-points to use along high-symmetry lines
+    """
 
     # For band-structures, unit cells should be in the standardized format
     pre_standardize_structure = True
