+++ src/simmate/calculators/vasp/tasks/relaxation/__init__.py
@@ -5,8 +5,13 @@ from .quality_01 import Quality01Relaxation
 from .quality_02 import Quality02Relaxation
 from .quality_03 import Quality03Relaxation
 from .quality_04 import Quality04Relaxation
+
 from .mit import MITRelaxation
+
 from .materials_project import MatProjRelaxation
+from .materials_project_hse import MatProjHSERelaxation
+from .materials_project_metal import MatProjMetalRelaxation
+
 from .materials_virtual_lab_endpoint import (
     MatVirtualLabCINEBEndpointRelaxation as NEBEndpointRelaxation,
 )
