### 2022-04-28 fb06f14
:cow: continue converting pymatgen sets
