+++ src/simmate/calculators/vasp/tasks/elastic/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .materials_virtual_lab import MatVirtualLabElastic
