+++ src/simmate/calculators/vasp/tasks/test/test_pymatgen_sets.py
@@ -5,6 +5,7 @@ import pytest
 
 from simmate.conftest import make_dummy_files
 from simmate.calculators.vasp.inputs import Incar, Potcar
+from simmate.toolkit.diffusion import MigrationImages
 
 # -----------------------------------------------------------------------------
 # Simmate tasks and their corresponding pymatgen sets are paired below
@@ -53,6 +54,17 @@ from simmate.calculators.vasp.tasks.relaxation import (
 )
 from pymatgen.io.vasp.sets import MVLGBSet
 
+from simmate.calculators.vasp.tasks.nudged_elastic_band import MITNudgedElasticBand
+from pymatgen.io.vasp.sets import MITNEBSet
+
+from simmate.calculators.vasp.tasks.nudged_elastic_band import (
+    MatVirtualLabClimbingImageNudgedElasticBand,
+)
+from pymatgen.analysis.diffusion.neb.io import MVLCINEBSet
+
+from simmate.calculators.vasp.tasks.relaxation import NEBEndpointRelaxation
+from pymatgen.analysis.diffusion.neb.io import MVLCINEBEndPointSet
+
 # -----------------------------------------------------------------------------
 
 MD_KWARGS = {
@@ -72,6 +84,7 @@ MD_KWARGS = {
         (MatProjStaticEnergy, MPStaticSet, {}),
         (MatProjHSERelaxation, MPHSERelaxSet, {}),
         (MatProjMetalRelaxation, MPMetalRelaxSet, {}),
+        (NEBEndpointRelaxation, MVLCINEBEndPointSet, {}),
         (MITDynamics, MITMDSet, MD_KWARGS),
         (MatProjDynamics, MPMDSet, MD_KWARGS),
         (MatProjHSEBandStructure, MPHSEBSSet, {"mode": "line"}),
@@ -115,6 +128,75 @@ def test_pymatgen_input_sets(
     # write both inputs
     simmate_task().setup(structure=structure, directory=tmpdir)
     pymatgen_set(structure, **pymatgen_kwargs).incar.write_file(incar_pymatgen_name)
+
+    # load incar
+    incar_simmate = Incar.from_file(incar_simmate_name)
+    incar_pymatgen = Incar.from_file(incar_pymatgen_name)
+
+    # compare
+    diff = incar_simmate.compare_incars(incar_pymatgen)["Different"]
+    # ignore kpoints input for now
+    diff.pop("KSPACING", None)
+
+    # in EDIFF__per_atom, there are sometimes rounding issues.
+    # We say there is no difference if they are within 100% error. This may
+    # seem like a large error, but it is meant to encompass differences like
+    # {'EDIFF': (4e-06, 1e-05)} where pymatgen rounded significantly. There
+    # really isn't an issue unless EDIFF is off by a factor of 10 (1000%)
+    if "EDIFF" in diff and "EDIFF__per_atom" in simmate_task.incar:
+        ediff_simmate, ediff_pymatgen = diff["EDIFF"]
+        percent_diff = (ediff_pymatgen - ediff_simmate) / ediff_pymatgen
+        if abs(percent_diff) < 1:
+            diff.pop("EDIFF")
+
+    # ensure incars are the same (there are no differences)
+    assert diff == {}
+
+
+@pytest.mark.pymatgen
+@pytest.mark.parametrize(
+    "simmate_task, pymatgen_set, pymatgen_kwargs",
+    [
+        (MITNudgedElasticBand, MITNEBSet, {}),
+        (MatVirtualLabClimbingImageNudgedElasticBand, MVLCINEBSet, {}),
+    ],
+)
+def test_pymatgen_input_sets_neb(
+    structure,
+    tmpdir,
+    mocker,
+    simmate_task,
+    pymatgen_set,
+    pymatgen_kwargs,
+):
+    """
[…]
