+++ src/simmate/calculators/vasp/tasks/nudged_elastic_band/__init__.py
@@ -2,5 +2,5 @@
 
 from .base import VaspNudgedElasticBandTask
 
-from .materials_virtual_lab_ci import MatVirtualLabCINEB
+from .materials_virtual_lab_ci import MatVirtualLabClimbingImageNudgedElasticBand
 from .mit import MITNudgedElasticBand
