### 2022-04-29 6dce769
:snake: fix anaconda links
test
