+++ src/simmate/calculators/vasp/database/relaxation.py
@@ -26,10 +26,10 @@ MITRelaxation, MITIonicStep = Relaxation.create_subclasses("MIT", module=__name_
     MatProjIonicStep,
 ) = Relaxation.create_subclasses("MatProj", module=__name__)
 
-NEBEndpointRelaxation, NEBEndpointIonicStep = Relaxation.create_subclasses(
-    "NEBEndpoint",
-    module=__name__,
-)
+(
+    MatVirtualLabCINEBEndpointRelaxation,
+    MatVirtualLabCINEBEndpointIonicStep,
+) = Relaxation.create_subclasses("NEBEndpoint", module=__name__)
 
 Quality00Relaxation, Quality00IonicStep = Relaxation.create_subclasses(
     "Quality00",
