+++ src/simmate/calculators/vasp/inputs/incar.py
@@ -1,6 +1,18 @@
 # -*- coding: utf-8 -*-
 
-import math
+from typing import Callable
+
+from simmate.calculators.vasp.inputs.incar_modifiers import (
+    keyword_modifier_density,
+    keyword_modifier_density_a,
+    keyword_modifier_density_b,
+    keyword_modifier_density_c,
+    keyword_modifier_per_atom,
+    keyword_modifier_smart_magmom,
+    keyword_modifier_smart_lmaxmix,
+    keyword_modifier_smart_ldau,
+    keyword_modifier_smart_ismear,
+)
 
 
 class Incar(dict):
@@ -452,285 +464,56 @@ class Incar(dict):
         # now return the new Incar
         return Incar(**new_parameters)
 
-    # ------------------------------
-    # All methods below are modifiers that users can apply. I may move these to
-    # a separate class (e.g. KeywordModifier).
+    @classmethod
+    def add_keyword_modifer(cls, keyword_modifer: Callable):
+        """
+        Dynamically sets a keyword_modifier function to the Incar class. When
+        loaded onto the Incar class, you can then use the modifiers name to
+        set a parameter based on the input structure.
 
-    @staticmethod
-    def keyword_modifier_density(structure, density):
-        """
-        The __density modifier means the user wants a specific density. They
-        provide this density in per-angstrom^3 units and we return the
-        structure-specific count that gives this density.
-        For example, density=10 and a structure lattice that volume of 5,
-        then this returns value=10*5=50.
-        """
-        # VASP expect integers for a lot of these values, so we round up
-        return math.ceil(structure.lattice.volume * density)
+        Names of modifiers MUST start with "keyword_modifier_" and
+        have unique endings (i.e. you can't have 2 "per_atom" modifiers).
 
-    @staticmethod
-    def keyword_modifier_density_a(structure, density):
-        """
-        The __density_a modifier means the user wants a specific density along
-        the A lattice vector. They provide this density in per-angstrom units
-        and we return the structure-specific count that gives this density.
-        For example, density=10 and a structure lattice that A vector of 5,
-        then this returns value=10*5=50.
-        """
-        # VASP expect integers for a lot of these values, so we round up
-        return math.ceil(structure.lattice.a * density)
+        For example, in order to allow settings like `ENCUT__per_atom`, we use
+        a `keyword_modifier_per_atom` function and add it to this class:
 
-    @staticmethod
-    def keyword_modifier_density_b(structure, density):
-        return math.ceil(structure.lattice.b * density)
+        ``` python
 
-    @staticmethod
-    def keyword_modifier_density_c(structure, density):
-        return math.ceil(structure.lattice.c * density)
+        # create your modifier logic as a function
+        def keyword_modifier_double_per_atom(structure, per_atom_value):
+            return per_atom_value * structure.num_sites * 2
 
-    @staticmethod
-    def keyword_modifier_per_atom(structure, per_atom_value):
-        """
-        The __per_atom modifier means the user wants a specific value per atom
-        in the unit cell. For example, EDIFF__per_atom=1e-5 and a structure
-        with 50 sites in it would return a value of 1e-5*50=50.
-        """
-        # VASP expect integers for a lot of these values, so we round up
-        return per_atom_value * structure.num_sites
+        # add it to the Incar class to make it active
+        Incar.add_keyword_modifer(keyword_modifier_double_per_atom)
 
-    @staticmethod
-    def keyword_modifier_smart_magmom(structure, override_values):
-        """
-        The __smart_magmom modifier goes through a series of checks to decide
-        what to set the MAGMOM as for VASP. In order of priority they are...
[…]
