+++ src/simmate/calculators/vasp/inputs/incar_modifiers.py
@@ -1,0 +1,251 @@
+# -*- coding: utf-8 -*-
+
+import math
+
+from simmate.toolkit import Structure
+
+# TODO: consider making modifiers a class. For example:
+#
+# class PerAtom(KeywordModifier):
+#     tag_name = "__per_atom"
+#     @staticmethod
+#     def apply(structure, per_atom_value):
+#         return per_atom_value * structure.num_sites
+
+
+def keyword_modifier_density(structure: Structure, density: float):
+    """
+    The __density modifier means the user wants a specific density. They
+    provide this density in per-angstrom^3 units and we return the
+    structure-specific count that gives this density.
+    For example, density=10 and a structure lattice that volume of 5,
+    then this returns value=10*5=50.
+    """
+    # VASP expect integers for a lot of these values, so we round up
+    return math.ceil(structure.lattice.volume * density)
+
+
+def keyword_modifier_density_a(structure: Structure, density: float):
+    """
+    The __density_a modifier means the user wants a specific density along
+    the A lattice vector. They provide this density in per-angstrom units
+    and we return the structure-specific count that gives this density.
+    For example, density=10 and a structure lattice that A vector of 5,
+    then this returns value=10*5=50.
+    """
+    # VASP expect integers for a lot of these values, so we round up
+    return math.ceil(structure.lattice.a * density)
+
+
+def keyword_modifier_density_b(structure: Structure, density: float):
+    return math.ceil(structure.lattice.b * density)
+
+
+def keyword_modifier_density_c(structure: Structure, density: float):
+    return math.ceil(structure.lattice.c * density)
+
+
+def keyword_modifier_per_atom(structure: Structure, per_atom_value: float):
+    """
+    The __per_atom modifier means the user wants a specific value per atom
+    in the unit cell. For example, EDIFF__per_atom=1e-5 and a structure
+    with 50 sites in it would return a value of 1e-5*50=50.
+    """
+    return per_atom_value * structure.num_sites
+
+
+def keyword_modifier_smart_magmom(structure: Structure, override_values: dict):
+    """
+    The __smart_magmom modifier goes through a series of checks to decide
+    what to set the MAGMOM as for VASP. In order of priority they are...
+        (1) the magmom property attached to each site in the structure object
+        (2) the spin property attached to the site's specie
+        (3) a value provided explicitly (e.g. {"Co": 0.5})
+        (4) a value of 0.6
+    """
+
+    # grab the default MAGMOM supplied, or use VASP's default of 1 otherwise
+    default_value = override_values.get("default", 1)
+
+    # we go through each site in the structure and decide what to set the
+    # MAGMOM for each. This allows even different sites of the same
+    # element to have their own MAGMOM
+    magnetic_moments = []
+    for site in structure:
+        # if the structure object has magmom-decorated sites, we use that
+        # as our first priority
+        if hasattr(site, "magmom"):
+            magnetic_moments.append(site.magmom)
+        # next we check if the site's specie has a spin
+        elif hasattr(site.specie, "spin"):
+            magnetic_moments.append(site.specie.spin)
+        # we then look at the override dictionary if there was one provided.
+        # If note, we use 0.6 as a default.
+        else:
+            magnetic_moment = override_values.get(site.specie.symbol, default_value)
+            magnetic_moments.append(magnetic_moment)
+
+    # This feature is in pymatgen, but I haven't added it here yet.
+    # if self.constrain_total_magmom:
+    #     nupdown = sum([mag if abs(mag) > 0.6 else 0 for mag in incar["MAGMOM"]])
[…]
