### 2022-05-06 #4364 @danielskatz
FYI, @editorialbot commands need to be the first thing in a command
