### 2022-05-06 #4364 @jacksund
@rkurchin @bocklund @utf Thank you for volunteering to edit and review our package! We're excited for the feedback.

I just wanted to place two things upfront to help with the review.

First, our test suite is currently at 84% coverage, and here's the report:
[htmlcov_2022-05-06.zip](https://github.com/openjournals/joss-reviews/files/8641898/htmlcov_2022-05-06.zip)

Second, I just wanted to repeat my comment from the pre-review thread:
We modeled our paper after the [example one](https://joss.theoj.org/papers/10.21105/joss.00388) provided in the docs, where the "Statement of Need" and "Summary" sections were combined. If that is no longer allowed, we can change this though!
