### 2022-05-21 3c5e1ca
:beetle: fix spelling of modifier in tests and usage
