+++ .github/workflows/lint-and-test.yml
@@ -20,19 +20,32 @@ on:
     - cron:  '30 5,17 * * *'
 
 jobs:
-  build-linux:
-    runs-on: ubuntu-latest
+  lint-and-test:
+
+    # Establishes the combination of all OS and python versions
+    # for us to test. The "steps" below will be ran for each
+    # of these combinations. 
+    name: Test (${{ matrix.python-version }}, ${{ matrix.os }})
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: ["ubuntu-latest", "macos-latest", "windows-latest"]
+        python-version: ["3.10"]
+
     steps:
 
-    - uses: actions/checkout@v2
+    - name: Initial setup
+      uses: actions/checkout@v2
 
-    - uses: psf/black@stable
+    - name: Check code format with Black
+      uses: psf/black@stable
 
     - name: Create conda env and install deps
       uses: conda-incubator/setup-miniconda@v2
       with:
         environment-file: tutorials/Guides_for_contributors/environment.yaml
-        python-version: "3.10"
+        python-version: ${{ matrix.python-version }}
         auto-activate-base: false
         channels: conda-forge
         activate-environment: simmate_dev
