### 2022-05-22 5e07142
update log with CI fixes
