### 2022-05-22 a258d6f
:robot: update CI w. pytest<7.1 & conda action (#167)
* :robot: use third-party conda action

* :bug: fix 3.10 with quotes

* confirm conda env

* reformat for reading

* test mac and windows

* limit pytest to 7.0

* update log with CI fixes
