### 2022-05-24 bccf8e2
Fix benchmarks documetnation typo
