### 2022-05-31 a9eccf1
:dizzy: condense parameter serialization
