+++ src/simmate/workflow_engine/workflow.py
@@ -3,6 +3,7 @@
 import json
 import cloudpickle
 import yaml
+from typing import List, Any, Union
 
 # note: extra modules are imported from prefect for convenience imports elsewhere
 import prefect
@@ -16,9 +17,10 @@ from prefect.backend import FlowRunView, FlowView
 # in order to support "mocking" this function in unit tests
 from prefect.backend import flow_run as flow_run_module
 
-from typing import List
-from simmate.workflow_engine.supervised_staged_shell_task import S3Task
+from simmate.toolkit import Structure
+from simmate.toolkit.diffusion import MigrationHop
 from simmate.database.base_data_types import DatabaseTable, Calculation
+from simmate.workflow_engine import S3Task
 
 
 class Workflow(PrefectFlow):
@@ -28,7 +30,8 @@ class Workflow(PrefectFlow):
     method, which allows us to register a calculation to a database table before
     we submit the workflow to Prefect Cloud.
 
-    To learn how to use this class, see [prefect.core.flow.Flow](https://docs.prefect.io/api/latest/core/flow.html#flow-2)
+    To learn how to use this class, see
+    [prefect.core.flow.Flow](https://docs.prefect.io/api/latest/core/flow.html#flow-2)
     """
 
     result_task: Task = None
@@ -124,6 +127,12 @@ class Workflow(PrefectFlow):
         # use yaml to make the printout pretty (no quotes and separate lines)
         print(yaml.dump(self.parameter_names))
 
+    def run(self, **kwargs: Any) -> Union["prefect.engine.state.State", None]:
+        # Simply deserializes inputs before calling the normal workflow.run
+        # method. We leave the docstring empty for inheritance.
+        kwargs_cleaned = self._deserialize_parameters(**kwargs)
+        return super().run(**kwargs_cleaned)
+
     def run_cloud(
         self,
         labels: List[str] = [],
@@ -174,8 +183,14 @@ class Workflow(PrefectFlow):
             project_name=self.project_name,
         )
 
-        # Serialize all of the parameters so they can be properly submitted
-        parameters_serialized = self._serialize_parameters(kwargs)
+        # Deserialize and then Serialize all of the parameters so they can be properly submitted
+        parameters_deserialized = self._deserialize_parameters(**kwargs)
+        parameters_serialized = self._serialize_parameters(**parameters_deserialized)
+        # NOTE: This may seem odd to deserialize right before serializing in the
+        # next line, but this is done to ensure parameters that accept file names
+        # are submitted properly. i.e -- We don't want to submit to a cluster and
+        # have the job fail because it doesn't have access to the file. Having
+        # these lines back-to-back ensures we submit with all necessary data.
 
         # Now we submit the workflow
         logger.info(f"Creating flow run for {self.name}...") if logger else None
@@ -249,7 +264,95 @@ class Workflow(PrefectFlow):
         return flow_run.get_latest()
 
     @staticmethod
-    def _serialize_parameters(parameters: dict) -> dict:
+    def _deserialize_parameters(**parameters) -> dict:
+        """
+        Converts input parameters to proper python objects. This handles the
+        variety of formats accepted by Simmate.
+
+        For example, a common input parameter for workflows is "structure", which
+        can be provided a number of ways:
+            - a filename
+            - a json string
+            - a dictionary pointing to a database entry
+            - a toolkit Structure object
+            - etc...
+        Even though all of these inputs are accepted, `workflow.run` always expects
+        python objects, so this utility converts the input to a toolkit Structure
+        object.
+
+        This method should not be called directly as it is used within higher-level
+        methods such as run() and run_cloud().
+        """
+
+        # we don't want to pass arguments like command=None or structure=None if the
+        # user didn't provide this input parameter. Instead, we want the workflow to
+        # use its own default value. To do this, we first check if the parameter
[…]
