### 2022-06-02 aa84a76
:radioactive: refactor load_input_and_register task
