+++ src/simmate/workflow_engine/common_tasks/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from .load_input_and_register import LoadInputAndRegister
+from .load_input_and_register import load_input_and_register
 from .save_result import SaveOutputTask
 from .load_nested_calculation import LoadNestedCalculationTask
 from .save_nested_calculation import SaveNestedCalculationTask
