+++ src/simmate/calculators/vasp/workflows/dynamics.py
@@ -5,22 +5,14 @@ from simmate.workflow_engine.workflow import (
     Parameter,
     ModuleStorage,
 )
-from simmate.workflow_engine.common_tasks import LoadInputAndRegister, SaveOutputTask
+from simmate.workflow_engine.common_tasks import load_input_and_register, SaveOutputTask
 from simmate.calculators.vasp.tasks.dynamics import MITDynamics
 from simmate.calculators.vasp.database.dynamics import MITDynamicsRun
 
-WORKFLOW_NAME = "dynamics/mit"
-
 s3task_obj = MITDynamics()
-load_input_and_register = LoadInputAndRegister(MITDynamicsRun)
-load_input_and_register = LoadInputAndRegister(
-    workflow_name=WORKFLOW_NAME,
-    input_obj_name="structure",
-    calculation_table=MITDynamicsRun,
-)
 save_results = SaveOutputTask(MITDynamicsRun)
 
-with Workflow(WORKFLOW_NAME) as workflow:
+with Workflow("dynamics/mit") as workflow:
     structure = Parameter("structure")
     command = Parameter("command", default="vasp_std > vasp.out")
     source = Parameter("source", default=None)
@@ -33,8 +25,8 @@ with Workflow(WORKFLOW_NAME) as workflow:
     time_step = Parameter("time_step", default=2)
     nsteps = Parameter("nsteps", default=10000)
 
-    structure_toolkit, directory_cleaned = load_input_and_register(
-        input_obj=structure,
+    parameters_cleaned = load_input_and_register(
+        structure=structure,
         source=source,
         directory=directory,
         command=command,
@@ -44,14 +36,15 @@ with Workflow(WORKFLOW_NAME) as workflow:
         time_step=time_step,
         nsteps=nsteps,
     )
+
     output = s3task_obj(
-        structure=structure_toolkit,
-        command=command,
-        directory=directory_cleaned,
-        temperature_start=temperature_start,
-        temperature_end=temperature_end,
-        time_step=time_step,
-        nsteps=nsteps,
+        structure=parameters_cleaned["structure"],
+        command=parameters_cleaned["command"],
+        directory=parameters_cleaned["directory"],
+        temperature_start=parameters_cleaned["temperature_start"],
+        temperature_end=parameters_cleaned["temperature_end"],
+        time_step=parameters_cleaned["time_step"],
+        nsteps=parameters_cleaned["nsteps"],
     )
     calculation_id = save_results(output=output)
 

[…]
