### 2022-06-02 4fef7cf
:biohazard: refactor save_result task
