+++ src/simmate/workflow_engine/common_tasks/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from .load_input_and_register import load_input_and_register
-from .save_result import SaveOutputTask
+from .save_result import save_result
 from .load_nested_calculation import LoadNestedCalculationTask
 from .save_nested_calculation import SaveNestedCalculationTask
 from .parse_multi_command import parse_multi_command
