### 2022-06-02 81f1a4c
:beetle: fix serialization of metadata
