### 2022-06-02 be269ed
:recycle: refactor register_kwargs and customized flows
