+++ src/simmate/calculators/vasp/workflows/density_of_states.py
@@ -14,6 +14,6 @@ workflow = s3task_to_workflow(
     project_name="Simmate-Density-of-States",
     s3task=MatProjDensityOfStatesTask,
     calculation_table=MatProjDensityOfStatesResults,
-    register_kwargs=["prefect_flow_run_id", "structure", "source"],
+    register_kwargs=["structure", "source"],
     description_doc_short="uses Materials Project settings (PBE)",
 )
