+++ src/simmate/calculators/vasp/workflows/relaxation/materials_project.py
@@ -14,6 +14,6 @@ workflow = s3task_to_workflow(
     project_name="Simmate-Relaxation",
     s3task=MPRelaxationTask,
     calculation_table=MPRelaxationResults,
-    register_kwargs=["prefect_flow_run_id", "structure", "source"],
+    register_kwargs=["structure", "source"],
     description_doc_short="uses Materials Project settings",
 )
