+++ src/simmate/calculators/vasp/workflows/nudged_elastic_band/all_paths.py
@@ -191,7 +191,6 @@ with Workflow("diffusion/all-paths") as workflow:
 workflow.storage = ModuleStorage(__name__)
 workflow.project_name = "Simmate-Diffusion"
 # workflow.calculation_table = MITDiffusionAnalysis  # not implemented yet
-# workflow.register_kwargs = ["prefect_flow_run_id"]
 workflow.result_table = MITDiffusionAnalysis
 workflow.s3tasks = [
     relaxation_mit_workflow.s3task,
