+++ src/simmate/calculators/vasp/workflows/nudged_elastic_band/single_path.py
@@ -111,7 +111,6 @@ with Workflow("diffusion/single-path") as workflow:
 workflow.storage = ModuleStorage(__name__)
 workflow.project_name = "Simmate-Diffusion"
 # workflow.calculation_table = MITDiffusionAnalysis  # not implemented yet
-# workflow.register_kwargs = ["prefect_flow_run_id"]
 workflow.result_table = MITDiffusionAnalysis
 workflow.s3tasks = [
     relaxation_neb_endpoint_workflow.s3task,
