+++ src/simmate/calculators/vasp/workflows/relaxation/quality_01.py
@@ -14,6 +14,6 @@ workflow = s3task_to_workflow(
     project_name="Simmate-Relaxation",
     s3task=Quality01RelaxationTask,
     calculation_table=Quality01RelaxationResults,
-    register_kwargs=["prefect_flow_run_id", "structure", "source"],
+    register_kwargs=["structure", "source"],
     description_doc_short="low-quality settings meant for highly unreasonable structures",
 )
