### 2022-06-09 8d28965
:cloud_with_lightning: add electronic structure workflow
