+++ src/simmate/workflow_engine/common_tasks/load_input_and_register.py
@@ -11,6 +11,7 @@ from prefect import task
 from simmate.toolkit import Structure
 from simmate.toolkit.diffusion import MigrationHop, MigrationImages
 from simmate.utilities import get_directory
+from simmate.workflow_engine import Workflow
 
 # OPTIMIZE: consider splitting this task into load_structure, load_directory,
 # and register_calc so that our flow_visualize looks cleaner
@@ -74,7 +75,13 @@ def load_input_and_register(register_run=True, **parameters: Any) -> dict:
     if not workflow_name:
         raise Exception("Unknown workflow")
 
-    workflow = get_workflow(workflow_name)
+    # BUG: I have to do this in order to allow workflows that are from the
+    # simmate.workflows module. There should be a better way to handle user
+    # created workflows.
+    try:
+        workflow = get_workflow(workflow_name)
+    except:
+        workflow = None
 
     prefect_flow_run_id = prefect.context.flow_run_id
 
@@ -196,12 +203,11 @@ def load_input_and_register(register_run=True, **parameters: Any) -> dict:
         # the "fabric" python package (uses ssh). I'd also need to store
         # filesystem names (e.g. "WarWulf") to know where to connect.
 
-    if "directory" in workflow.parameter_names:
-        # SPECIAL CASE for customized flows
-        if "workflow_base" not in parameters_cleaned:
-            parameters_cleaned["directory"] = directory_cleaned
-        else:
-            parameters_cleaned["input_parameters"]["directory"] = directory_cleaned
+    # SPECIAL CASE for customized flows
+    if "workflow_base" not in parameters_cleaned:
+        parameters_cleaned["directory"] = directory_cleaned
+    else:
+        parameters_cleaned["input_parameters"]["directory"] = directory_cleaned
 
     # ---------------------------------------------------------------------
 
@@ -224,8 +230,7 @@ def load_input_and_register(register_run=True, **parameters: Any) -> dict:
     else:
         source_cleaned = None
 
-    if "source" in workflow.parameter_names:
-        parameters_cleaned["source"] = source_cleaned
+    parameters_cleaned["source"] = source_cleaned
 
     # ---------------------------------------------------------------------
 
@@ -235,7 +240,7 @@ def load_input_and_register(register_run=True, **parameters: Any) -> dict:
     # don't store calculation information bc the flow is just a quick python
     # analysis.
 
-    if register_run and workflow.calculation_table:
+    if register_run and workflow and workflow.calculation_table:
 
         # grab the registration kwargs from the parameters provided
         register_kwargs = {
@@ -266,7 +271,7 @@ def load_input_and_register(register_run=True, **parameters: Any) -> dict:
     # structure data instead of a filename in the metadata.
     # SPECIAL CASE: "if ..." used to catch customized workflows
     parameters_serialized = (
-        workflow._serialize_parameters(**parameters_cleaned)
+        Workflow._serialize_parameters(**parameters_cleaned)
         if "workflow_base" not in parameters_cleaned
         else parameters
     )
@@ -275,7 +280,7 @@ def load_input_and_register(register_run=True, **parameters: Any) -> dict:
     # workflow run. This allows future users to reproduce the results if
     # desired -- and it also allows us to load old results into a database.
     input_summary = dict(
-        workflow_name=workflow.name,
+        workflow_name=workflow.name if workflow else "non-module-flow",
         # this ID is ingored as an input but needed for loading past data
         prefect_flow_run_id=prefect_flow_run_id,
         **parameters_serialized,
