### 2022-06-09 e9e4971
:battery: register electronic wf
