### 2022-06-09 0cfc449
remove source if none
