### 2022-06-09 daf2f5f
add debug prints
