### 2022-06-12 eaab1a4
move file gen into badelf task
