### 2022-06-12 dcb62dc
check for empty files in bader task
