+++ src/simmate/workflows/population_analysis.py
@@ -9,4 +9,5 @@ Workflows for predicting oxidation states within a material
 from simmate.calculators.vasp.workflows.population_analysis import (
     bader_matproj_workflow,
     elf_matproj_workflow,
+    badelf_matproj_workflow,
 )
