+++ src/simmate/calculators/vasp/workflows/population_analysis/__init__.py
@@ -2,3 +2,4 @@
 
 from .bader import workflow as bader_matproj_workflow
 from .elf import workflow as elf_matproj_workflow
+from .badelf import workflow as badelf_matproj_workflow
