+++ src/simmate/workflows/population_analysis.py
@@ -10,4 +10,6 @@ from simmate.calculators.vasp.workflows.population_analysis import (
     bader_matproj_workflow,
     elf_matproj_workflow,
     badelf_matproj_workflow,
+    prebader_matproj_workflow,
+    prebadelf_matproj_workflow,
 )
