### 2022-06-14 865f4e7
:cloud_with_lightning: add electronic and badelf workflows (#174)
* add electronic structure workflow

* :battery: register electronic wf

* fix source cleaning bug

* remove source if none

* add debug prints

* add explicit null source

* remove debug prints

* begin adding badelf flow

* fix fft grid in badelf

* move file gen into badelf task

* check for empty files in bader task

* test badelf with low settings

* add badelf flow

* register pre bader flows

* fix db typo

* increase badelf quality

* fix loading

* fix mapping bug for badelf

* check for version update
