+++ src/simmate/calculators/vasp/tasks/population_analysis/__init__.py
@@ -2,4 +2,4 @@
 
 from .pre_bader import MatProjPreBader
 from .elf import MatProjELF
-from .elf_bader import MatProjPreBaderELF
+from .pre_badelf import MatProjPreBaderELF
