### 2022-06-15 #4364 @bocklund
@rkurchin is an explicit "Statement of Need" section required for the paper? ([question here](https://github.com/openjournals/joss-reviews/issues/4364#issuecomment-1119762968)).
