### 2022-06-15 jacksund/simmate#176 @bocklund
JOSS software paper feedback
As part of my review in https://github.com/openjournals/joss-reviews/issues/4364, I have two comments on the software paper:

1. Incorporate "Statement of Need", if required https://github.com/openjournals/joss-reviews/issues/4364#issuecomment-1156705239
2. I think it would be nice to include a few additional details about what makes Simmate different from the other tools that you mentioned in the paper. I basically have in mind that you can borrow some of the content that you have in the README already under the "What is Simmate" heading and what is in the [comparison document](https://github.com/jacksund/simmate/tree/main/benchmarks)

Otherwise, it looks great 😄
