### 2022-06-15 jacksund/simmate#176 @jacksund
Awesome! Thank you for reading through the paper.

For 1, I'll wait and see wait the JOSS editors say. Shouldn't be a big switch either way.

For 2, thanks for the recommendation! I'll add some more details that compare features and also cite the main inspirations for our modules.
