### 2022-06-15 #4364 @rkurchin
Thanks for checking. @jacksund, if you could split those up I think that would be preferable. While the rendered version of that paper (from 2017) has them combined, if you look at the [updated source in the docs](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography), it has separate headers, so I think that's what's recommended currently.
