+++ paper.bib
@@ -175,3 +175,49 @@
     url = {https://doi.org/10.1016/j.commatsci.2015.09.013},
     author = {{Pizzi}, G. and {Cepellotti}, A. and {Sabatini}, R. and {Marzari}, N. and {Kozinsky}, B.},
 }
+
+@misc{emmet,
+    author = {{EMMET Collaboration}},
+    title = {EMMET: Be a master builder of databases of material properties. Avoid the Kragle.},
+    year = {2022},
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    url = {https://github.com/materialsproject/emmet},
+}
+
+
+@article{atomate,
+    title = {Atomate: A high-level interface to generate, execute, and analyze computational materials science workflows},
+    author = {{Mathew}, K. and {Montoya}, J.H. and {Faghaninia}, A. and {Dwarakanath}, S. and {Aykol}, M. and {Tang}, H. and {Chu}, I. and {Smidt}, T. and {Bocklund}, B. and {Horton}, M. and {Dagdelen}, J. and {Wood}, B. and {Liu}, Z. and {Neaton}, J. and {Ong}, S.~P. and {Persson}, K. and {Jain}, A.},
+    journal = {Computational Materials Science},
+    volume = {139},
+    issn = {0927-0256},
+    number = {140–152},
+    year = {2017},
+    doi = {10.1016/j.commatsci.2017.07.030},
+    url = {https://doi.org/10.1016/j.commatsci.2017.07.030},
+}
+
+
+@misc{mpcontrib,
+    author = {{MPContrib Collaboration}},
+    title = {MPContrib: Platform for materials scientists to contribute and disseminate their materials data through Materials Project},
+    year = {2022},
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    url = {https://github.com/materialsproject/MPContribs},
+}
+
+
+@article{fireworks,
+    author = {{Jain}, A. and {Ong}, S.~P. and {Chen}, W. and {Medasani}, B. and {Qu}, X. and {Kocher}, M. and {Brafman}, M. and {Petretto}, G. and {Rignanese}, G.~M. and {Hautier}, G. and {Gunter}, D. and {Persson}, K.~A.},
+    title = {FireWorks: a dynamic workflow system designed for high-throughput applications},
+    journal = {Concurrency and Computation: Practice and Experience},
+    volume = {27},
+    number = {17},
+    pages = {5037-5059},
+    doi = {10.1002/cpe.3505},
+    url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/cpe.3505},
+    year = {2015}
+}
+
