### 2022-06-20 b78257d
begin refactor for evolution search algo
