+++ src/simmate/toolkit/structure_prediction/evolution/search_engine.py
@@ -5,13 +5,13 @@ import time
 from simmate.toolkit import Composition
 
 from simmate.database import connect
-from simmate.database.workflow_results.evolution import (
+from simmate.database.workflow_results import (
     EvolutionarySearch as SearchDatatable,
     StructureSource as SourceDatatable,
 )
 
 import simmate.toolkit.creators as creation_module
-import simmate.toolkit.transformations as transform_module
+import simmate.toolkit.transformations.from_ase as transform_module
 from simmate.toolkit.validators.fingerprint.pcrystalnn import (
     PartialCrystalNNFingerprint,
 )
@@ -19,48 +19,23 @@ from simmate.toolkit.validators.fingerprint.pcrystalnn import (
 from typing import List, Union, Tuple
 from simmate.workflow_engine import Workflow
 
-# TODO:
-#
-# what if the workflow changes (e.g. starts with ML relax) but the
-# final table stays the same?
-#
-# chemical system instead of composition input? Allow 1,2,3D?
-#
-# I assume fitness_field="energy_per_atom" for now. Allow other fields from
-# the individuals_table, like "conductivity" or "electride character"
-#
-# Specifying a stop_condition class to allow custom ones?
-# I just assume the stop condition is either (1) the maximum allowed
-# calculated structures or (2) how long a given structure has been
-# the "best" one.
-#
-# add triggered_actions which can be things like "start_ml_potential" or
-# "update volume in sources" or "switch workflow to __"
-#
-# I assume we are using Prefect for workflow submission right now. Add support
-# for local execution or even dask.
-#
-# Add singleshot sources
-# "prototypes_aflow",
-# "substitution",
-# "third_party_structures",
-# "third_party_substitution",
-# "SomeOtherDatatable",  # e.g. best from table of lower quality calculations
-
 
 class SearchEngine:
     """
+    **WARNING** This search engine assumes you have properly configured
+    Prefect Cloud and a cloud database backend (e.g. Postgres). In the future,
+    we will accommodate local runs and other backends.
+
     This class is the entry point for predicting crystal structures with an
     evolutionary search algorithm.
 
     Once you initialize a SearchEngine with a composition (and any other
     optional parameters), you simply need to call the run() method to start
-    the search. Note that we've added this class to the shortcuts module for
-    convenience:
+    the search:
 
-    .. code-block:: python
+    ``` python
 
-        from simmate.shortcuts import SearchEngine
+        from simmate.toolkit.structure_prediction import SearchEngine
 
         search_engine = SearchEngine(
             composition="C4",
@@ -69,16 +44,19 @@ class SearchEngine:
         )
 
         search.run()
+    ```
 
     This class is only for creating (or continuing) searches. If you only want
     to view results, you should instead look directly at the database table for
     previously ran searches:
 
-    .. code-block:: python
+    ``` python
 
        from simmate.shortcuts import SearchResults
 
        search_results = SearchResults.objects.get(composition="Sr4 Si4 N8")
+
+    ```
 
     If you start a separate terminal from where a search is running, you can
     actually use this table to view results WHILE the search is still running!
@@ -86,29 +64,21 @@ class SearchEngine:
     You can also access the search results from your search_engine object. So
     once your search is finished or stopped, you can do:
 
-    .. code-block:: python
-
+    ``` python
        search_results = search_engine.search_datatable
+    ```
 
     To better understand how to view/analyze results, please read the documentation
     for the SearchResults class.
 
-    **WARNING** This search engine assumes you have properly configured
-    Prefect Cloud and a cloud database backend (e.g. Postgres). In the future,
-    we will accommodate local runs and other backends.
+    #### Alternative Codes & Softwares
 
-    Alternative Codes & Softwares
-    ----------------------------
[…]
