+++ src/simmate/workflow_engine/workflow.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 
+import os
+import shutil
 import json
 import cloudpickle
 import yaml
@@ -181,28 +183,14 @@ class Workflow(PrefectFlow):
             project_name=self.project_name,
         )
 
-        # -----------------------------------
-        # This part is unique to Simmate. Because we often want to save some info
-        # to our database even before the calculation starts/finishes, we do that
-        # here. An example is storing the structure and prefect id that we just
-        # submitted.
-        # BUG: Will there be a race condition here? What if the workflow finishes
-        # and tries writing to the databse before this is done?
-
-        # Deserialize and then Serialize all of the parameters so they can be properly submitted
-        from simmate.workflow_engine.common_tasks import load_input_and_register
-
-        parameters_deserialized = load_input_and_register.run(
-            workflow_name=self.name,
-            **kwargs,
-        )
-        parameters_serialized = self._serialize_parameters(**parameters_deserialized)
-        # NOTE: This may seem odd to deserialize right before serializing in the
-        # next line, but this is done to ensure parameters that accept file names
-        # are submitted properly. i.e -- We don't want to submit to a cluster and
-        # have the job fail because it doesn't have access to the file. Having
-        # these lines back-to-back ensures we submit with all necessary data.
-        # -----------------------------------
+        # Prefect does not properly deserialize objects that have
+        # as as_dict or to_dict method, so we use a custom method to do that here
+        parameters_serialized = self._serialize_parameters(**kwargs)
+        # BUG: What if we are submitting using a filename? We don't want to
+        # submit to a cluster and have the job fail because it doesn't have
+        # access to the file. One solution could be to deserialize right before
+        # serializing in the next line in order to ensure parameters that
+        # accept file names are submitted with all necessary data.
 
         # Now we submit the workflow
         logger.info(f"Creating flow run for {self.name}...") if logger else None
@@ -217,10 +205,22 @@ class Workflow(PrefectFlow):
         run_url = client.get_cloud_url("flow-run", flow_run_id, as_user=False)
         logger.info(f"Created flow run: {run_url}") if logger else None
 
+        # -----------------------------------
+        # This part is unique to Simmate. Because we often want to save some info
+        # to our database even before the calculation starts/finishes, we do that
+        # here. An example is storing the structure and prefect id that we just
+        # submitted.
+        # BUG: Will there be a race condition here? What if the workflow finishes
+        # and tries writing to the databse before this is done?
+        parameters_cleaned = self._deserialize_parameters(**kwargs)
+        self._register_calculation(flow_run_id, parameters_cleaned)
+        # -----------------------------------
+
         # if we want to wait until the job is complete, we do that here
         if wait_for_run:
             flow_run_view = self.wait_for_flow_run(flow_run_id)
             # return flow_run_view # !!! Should I return this instead?
+
         # return the flow_run_id for the user
         return flow_run_id
 
@@ -239,8 +239,40 @@ class Workflow(PrefectFlow):
         for log in flow_run_module.watch_flow_run(flow_run_id):
             message = f"Flow {flow_run.name}: {log.message}"
             prefect.context.logger.log(log.level, message)
+
         # Return the final view of the flow run
         return flow_run.get_latest()
+
+    def _register_calculation(self, prefect_flow_run_id, parameters):
+        """
+        If the workflow is linked to a calculation table in the Simmate database,
+        this adds the flow run to the database.
+
+        Parameters passed should be deserialized and cleaned.
+
+        This method should not be called directly as it is used within the
+        `run_cloud` method and `load_input_and_register` task.
+        """
+
+        # grab the registration kwargs from the parameters provided
[…]
