### 2022-06-20 b92c132
remove as_user functionality
