### 2022-06-21 704ddac
misc updates to cli and staged flow
