+++ src/simmate/workflow_engine/workflow.py
@@ -1,7 +1,5 @@
 # -*- coding: utf-8 -*-
 
-import os
-import shutil
 import json
 import cloudpickle
 import yaml
