+++ src/simmate/file_converters/structure/database.py
@@ -4,30 +4,61 @@
 This module provides conversion between Simmate toolkit and database structures.
 
 Everything is built-in to the base toolkit/database classes, so you typically 
-convert between these two without ever loading this module directly.
-
-Recommended use:
+convert between these two without ever loading this module directly. Below are 
+some examples of how we recommend doing this.
+    
+EXAMPLE 1: Database object to Toolkit object
 ``` python
-# using MatProj as an example, but the same applies for all Structure tables
+# Grab an example database structure.
+# We use MatProj as an example, but the same applies for all Structure tables.
 from simmate.database.third_parties import MatProjStructure
-
-db_structure = MatProjStructure.objects.first()
+database_structure = MatProjStructure.objects.first()
 
 # convert to toolkit
-tk_structure = db_structure.to_toolkit()
+toolkit_structure = database_structure.to_toolkit()
 
-# convert back to database (as new row)
-db_structure2 = MatProjStructure.from_toolkit(structure=structure)
+# access the original database object and confirm we have the same thing
+database_structure_2 = toolkit_structure.database_object
+assert database_structure == database_structure_2
+```
+
+EXAMPLE 2: Toolkit object to Database object
+``` python
+# Grab an example toolkit structure.
+# We use a cif file, but you can load this structure from anywhere
+from simmate.toolkit import Structure
+toolkit_structure = Structure.from_file("cif")
+
+# Convert back to database (as new row)
+# We save this to the Materials Project Database (which you should NOT do), but
+# in practice, you should have your own independent table to do this.
+database_structure = MatProjStructure.from_toolkit(structure=toolkit_structure)
+database_structure.save()
+```
+
+EXAMPLE 3: Database metadata (dict) to Toolkit object
+``` python
+
+# Setup some database metadata
+example_metadata = {
+    "database_table": "MatProjStructure",
+    "database_id": "mp-123",
+}
+
+# Grab the structure directly as a toolkit object
+structure_toolkit = Structure.from_database_dict(example_metadata)
+
+# access the original database object
+database_structure = toolkit_structure.database_object
 ```
 """
 
 from django.utils.module_loading import import_string
 
-from simmate.database import connect
-from simmate.website.workflows import models as all_datatables
 from simmate.toolkit import Structure as ToolkitStructure
-
-# from simmate.database.base_data_types import Structure as DatabaseStructure
+from simmate.database import connect
+from simmate.database.base_data_types import Structure as DatabaseStructure
+from simmate.website.workflows import models as all_datatables
 
 
 class DatabaseAdapter:
@@ -44,24 +75,15 @@ class DatabaseAdapter:
     # def get_database(structure: ToolkitStructure) -> DatabaseStructure:
     #     return ...
 
-    # This method is not included as it is obtuse and promotes improper use.
-    # We keep this comment here for those inspecting the code for understanding.
-    #
-    # @staticmethod
-    # def get_toolkit(structure: DatabaseStructure) -> ToolkitStructure:
-    #     return structure.to_toolkit()
-
     @staticmethod
-    def get_toolkit_from_database_dict(structure: dict) -> ToolkitStructure:
+    def get_toolkit_from_database_dict(structure_dict: dict) -> ToolkitStructure:
         """
-        This is an experimental feature.
-
         Loads a structure from the Simmate database from a dictionary of
         metadata.
         """
 
         # start by loading the datbase table, which is given as a module path
-        datatable_str = structure["calculation_table"]
+        datatable_str = structure_dict["database_table"]
 
         # Import the datatable class -- how this is done depends on if it
         # is from a simmate supplied class or if the user supplied a full
@@ -75,43 +97,60 @@ class DatabaseAdapter:
[…]
