+++ src/simmate/calculators/vasp/workflows/relaxation/materials_project.py
@@ -13,7 +13,7 @@ workflow = s3task_to_workflow(
     module=__name__,
     project_name="Simmate-Relaxation",
     s3task=MPRelaxationTask,
-    calculation_table=MPRelaxationResults,
+    database_table=MPRelaxationResults,
     register_kwargs=["structure", "source"],
     description_doc_short="uses Materials Project settings",
 )

[…]
