### 2022-06-22 25ceb58
begin substitution util
