### 2022-06-23 d230739
move aflowprototype to third_party module
