+++ src/simmate/database/prototypes/README.md
@@ -1,2 +1,0 @@
-
-This module is still early in development and not ready for public use.
