+++ CHANGELOG.md
@@ -16,6 +16,7 @@ There is one key exception to the rules above -- and that is with `MAJOR`=0 rele
 > :bulb: For ongoing changes that have not been finalized/merged yet, view our [active pull-requests](https://github.com/jacksund/simmate/pulls) on github
 
 **Enhancements**
+- add `AflowPrototypes` to the `database.third_parties` module (only includes data distributed through pymatgen)
 - add new modules to `toolkit.structure_prediction` and `toolkit.creation`, including ones to provide `known`, `substitution`, and `prototype` based structures.
 - add `created_at` and `updated_at` columns to all database tables
 - check if there is a newer version of Simmate available and let the user know about the update
