### 2022-06-23 ec5ebac
refactor prototype creators
