### 2022-06-24 bba4e08
:see_no_evil: refactors for evo search alg (#179)
* begin refactor for evolution search algo

* run black

* remove as_user functionality

* misc updates to cli and staged flow

* switch to use of database_table

* begin substitution util

* move aflowprototype to third_party module

* fix doc typos

* refactor prototype creators

* add highlevel utility for prototypes
