+++ src/simmate/calculators/vasp/workflows/static_energy/materials_project.py
@@ -13,7 +13,7 @@ workflow = s3task_to_workflow(
     module=__name__,
     project_name="Simmate-Energy",
     s3task=MPStaticEnergyTask,
-    calculation_table=MPStaticEnergyResults,
+    database_table=MPStaticEnergyResults,
     register_kwargs=["structure", "source"],
     description_doc_short="uses Materials Project settings",
 )
