### 2022-06-25 jacksund/simmate#180 @bocklund
`simmate.visualization.structure.blender` missing `__init__.py`
## Describe the bug

After installing from conda-forge, and trying to start the server, I'm getting:

```python
Traceback (most recent call last):
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/site-packages/django/core/handlers/exception.py", line 47, in inner
    response = get_response(request)
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/site-packages/django/core/handlers/base.py", line 167, in _get_response
    callback, callback_args, callback_kwargs = self.resolve_request(request)
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/site-packages/django/core/handlers/base.py", line 290, in resolve_request
    resolver_match = resolver.resolve(request.path_info)
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/site-packages/django/urls/resolvers.py", line 590, in resolve
    for pattern in self.url_patterns:
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/site-packages/django/utils/functional.py", line 48, in __get__
    res = instance.__dict__[self.name] = self.func(instance)
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/site-packages/django/urls/resolvers.py", line 632, in url_patterns
    patterns = getattr(self.urlconf_module, "urlpatterns", self.urlconf_module)
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/site-packages/django/utils/functional.py", line 48, in __get__
    res = instance.__dict__[self.name] = self.func(instance)
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/site-packages/django/urls/resolvers.py", line 625, in urlconf_module
    return import_module(self.urlconf_name)
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/importlib/__init__.py", line 126, in import_module
  File "<frozen importlib._bootstrap>", line 1050, in _gcd_import
  File "<frozen importlib._bootstrap>", line 1027, in _find_and_load
  File "<frozen importlib._bootstrap>", line 1006, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 688, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 883, in exec_module
  File "<frozen importlib._bootstrap>", line 241, in _call_with_frames_removed
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/site-packages/simmate/website/core/urls.py", line 52, in <module>
    view=include(
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/site-packages/django/urls/conf.py", line 34, in include
    urlconf_module = import_module(urlconf_module)
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/importlib/__init__.py", line 126, in import_module
  File "<frozen importlib._bootstrap>", line 1050, in _gcd_import
  File "<frozen importlib._bootstrap>", line 1027, in _find_and_load
  File "<frozen importlib._bootstrap>", line 1006, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 688, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 883, in exec_module
  File "<frozen importlib._bootstrap>", line 241, in _call_with_frames_removed
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/site-packages/simmate/website/core_components/urls.py", line 5, in <module>
    from simmate.website.core_components import views
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/site-packages/simmate/website/core_components/views.py", line 11, in <module>
    from simmate.visualization.structure import make_blender_structure
  File "/Users/brandon/miniforge3/envs/joss-simmate/lib/python3.10/site-packages/simmate/visualization/structure/__init__.py", line 1, in <module>
    from .blender.base import make_blender_structure
ModuleNotFoundError: No module named 'simmate.visualization.structure.blender.base'
```

`simmate.visualization.structure.blender` seems to be missing an `__init__.py` and probably isn't being detected as a package by `setup.py`.

## To Reproduce

1. Install simmate from conda-forge, I ran `conda create -n joss-simmate simmate`
2. `simmate database reset`
3. `simmate run-server` - gives the traceback from the command line and an error page in the browser.

https://github.com/openjournals/joss-reviews/issues/4364
