### 2022-06-25 jacksund/simmate#180 @jacksund
For some reason, I can't reproduce this error, but the ` __init__.py` is definitely missing like you show. I'll update the conda-feedstock to test the many different imports to see if there are other places where this is an issue. Sorry you ran into this at such an early setup stage!
