+++ src/simmate/visualization/structure/blender/__init__.py
@@ -1,0 +1,1 @@
+from .base import make_blender_structure
