### 2022-06-25 jacksund/simmate#180 @jacksund
@bocklund I'm not able to reproduce this with anaconda, so the bug might only exist when using miniforge installs (which I've never tried before). When testing out the repo, I'd stick to anaconda if you can! That's how me and my team do installs, and it's what has been tested.
