### 2022-06-25 91ca848
:rainbow: prep v0.6.0 release (#182)
