+++ paper.md
@@ -27,11 +27,21 @@ bibliography: paper.bib
 
 # Summary
 
-Over the past decade, the automation of electronic structure codes has led to many large-scale initiatives for materials discovery, such as the Materials Project [@matproj], AFLOW [@aflow], OQMD [@oqmd], JARVIS [@jarvis], and others [@matcloud; @nomad; @aiida]. Each of these projects facilitate the creation and distribution of materials science data to the broader researcher community through databases, workflow libraries, and web interfaces. However, each software ecosystem (i.e. the collection of software used by a specific project) still possesses several pain-points for users attempting to implement new standards for computational research. Proper setup involves learning how (i) workflows are defined/orchestrated, (ii) how databases are built/accessed, and (iii) how website interfaces/APIs make results accessible to the community, where each component requires learning a new package and, more importantly, learning how that package integrates with others. As a result, it can be difficult to integrate several smaller packages when building a production-ready server and database for materials science research.
+Over the past decade, the automation of electronic structure codes has led to many large-scale initiatives for materials discovery, such as the Materials Project [@matproj], AFLOW [@aflow], OQMD [@oqmd], JARVIS [@jarvis], and others [@matcloud; @nomad; @aiida]. Each of these projects facilitate the creation and distribution of materials science data to the broader researcher community through databases, workflow libraries, and web interfaces. However, each software ecosystem (i.e. the collection of software used by a specific project) still possesses several pain-points for users attempting to implement new standards for computational research. Proper setup involves learning how (i) workflows are defined/orchestrated, (ii) how databases are built/accessed, and (iii) how website interfaces/APIs make results accessible to the community, where each component requires learning a new package and, more importantly, learning how that package integrates with others. As a result, it can be difficult to integrate several smaller packages when building a production-ready server and database for materials science research. To address this, we developed the Simulated Materials Ecosystem (Simmate).
 
-To address this, we developed the Simulated Materials Ecosystem (Simmate). Simmate strives to simplify the process for researchers who are setting up a full-featured server. For the purposes of beginners, we desired a framework that could run locally without requiring any additional setup. For the purposes of experts, we sought to facilitate scaling of calculations across any number of resources. Simmate accomplishes these by building on top of popular, well-established packages for workflow orchestration, database management, and materials science analysis. This includes integration with high-level, beginner-friendly packages such as Django [@django], Dask [@dask], Prefect [@prefect], and PyMatGen [@pymatgen]. The use of popular packages also lets Simmate users take advantage of these packages' large user communities, abundant guides, and robust coding standards, while Simmate handles the integration of these packages in the context of materials science. Because Simmate removes many obstacles to advanced computation, we anticipate that this code will be utilized by beginners and experts alike. Thus, our tutorials are written for researchers that have never used the command-line or python However, as users become comfortable, they can begin exploring underlying API and integrated packages for advanced features.
+# Statement of Need
 
[…]
