### 2022-06-25 jacksund/simmate#176 @jacksund
@bocklund I updated the manuscript using your feedback. Let me know if there are any other edits you'd like! I'll leave this issue open until you give the thumbs up :smile:
