### 2022-06-25 jacksund/simmate#180 @bocklund
Thanks for the quick fix!

FWIW, I also was able to reproduce this on another Mac with just conda installed, but I realized my steps to reproduce were incomplete:

1. Install simmate from conda-forge, I ran `conda create -n joss-simmate simmate`
2. `simmate database reset`
3. `simmate run-server` - gives the traceback from the command line and an error page in the browser.
4. **Visit the development server URL in the browser**
