### 2022-06-25 jacksund/simmate#180 @jacksund
ah yep, I'm seeing the error now. Haha sorry, I should have at least opened up the server. Not sure how this is sneaking through my unittests because I do test the server views...

Anyways, it looks to be working on v0.6.0 when I test it manually with pip install on Ubuntu. It can take up to 24thrs for the new version to be available on conda-forge.
