+++ src/simmate/calculators/vasp/configuration.md
@@ -1,0 +1,189 @@
+# VASP Installation
+
+Official guides can be found on the VASP wiki [here](https://www.vasp.at/wiki/index.php/Installing_VASP.6.X.X). For example, there are clear [guides](https://www.vasp.at/wiki/index.php/Personal_computer_installation) for installing VASP onto your personal computer, such as VASP 6 on Ubuntu 22.04.
+
+
+## VASP 5.4.4 on Ubuntu 22.04
+
+This guide is specifically for the Warren Lab as it requires build files that we share within the team. 
+
+To guarantee compatibility, we need to build all vasp dependencies by hand (for example, Ubuntu uses gcc v11 but vasp requires v9). For the Warren Lab, we have packaged everything in one zip file to make setup as simple as possible. Copy /media/synology/software/vasp/vasp.zip from WarWulf to your computer, such as your home directory (e.g. `/home/jacksund/`). This file is only 172.1MB, but will be over 9GB once we are finished installing vasp.
+
+Within the vasp directory, there are folders from (1) to (9) for each build step. These are the 9 programs that need to be built from source. Source files (*.tar.gz) are provided in each directory. Each folder also contains an "install" directory where the program will be installed. You'll also find another folder (usually the program name) that contains the unzipped contents of the *.tar.gz.
+
+Make sure the necessary build tools are installed and up to date:
+``` bash
+sudo apt update
+sudo apt upgrade
+sudo apt install gcc make m4 g++ gfortran
+``` 
+
+1. install `gmp`
+
+``` bash
+# NOTE: steps are effectively the same with other programs but
+# we only include comments on this first install
+
+# open the folder for the step we are on
+cd ~/vasp/01_gmp
+
+# unzip our build files
+tar xvzf *.tar.gz
+
+# switch into the directory before performing the rest of install steps
+cd gmp-6.2.1
+
+# update this command with the proper path + username
+./configure --prefix=/home/jacksund/vasp/01_gmp/install
+
+# PAUSE AND READ OUTPUT OF THIS COMMAND
+# Look for "Thread(s) per core and use" this in the next command
+lscpu
+
+# “2” here is based on our output from the previous command. This value will 
+# be used in building our other packages too 
+make -j 2
+make install
+make check # (optional) to confirm successful install
+```
+
+2. install `mpfr`
+``` bash
+cd ~/vasp/02_mpfr
+tar xvzf *.tar.gz
+cd mpfr-4.1.0
+./configure --prefix=/home/jacksund/vasp/02_mpfr/install --with-gmp=/home/jacksund/vasp/01_gmp/install
+make -j 2
+make install
+make check
+```
+
+3. install `mpc` 
+``` bash
+cd ~/vasp/03_mpc
+tar xvzf *.tar.gz
+cd mpc-1.2.1
+./configure --prefix=/home/jacksund/vasp/03_mpc/install --with-gmp=/home/jacksund/vasp/01_gmp/install --with-mpfr=/home/jacksund/vasp/02_mpfr/install
+make -j 2
+make install
+make check
+```
+ 
+4. install `gcc`
+``` bash
+cd ~/vasp/04_gcc
+tar xvzf *.tar.gz
+cd gcc-9.5.0
[…]
