### 2022-07-05 889e549
:books: add installation guide for vasp (#183)
* Create configuration.md

* Update CHANGELOG.md
