### 2022-07-05 8ffe951
add utilities for loading remote archives
