### 2022-07-05 e1a9986
add init text
