+++ src/simmate/toolkit/structure_prediction/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from .known import get_known_structures
+from .prototypes import get_structures_from_prototypes
+from .substitution import get_structures_from_substitution_of_known
