+++ src/simmate/command_line/database.py
@@ -2,7 +2,7 @@
 
 """
 This defines commands for managing your Simmate database. All commands are 
-accessible through the "simmate database" command.
+accessible through the `simmate database` command.
 """
 
 import click
@@ -31,12 +31,31 @@ def reset(confirm_delete):
             "Do you still want to continue?",
             abort=True,
         )
+
+    # if the user has a SQLite3 backend, ask if they'd like to use a prebuilt
+    # database to begin
+    from simmate.configuration.django.settings import DATABASES
+
+    using_sqlite = DATABASES["default"]["ENGINE"] == "django.db.backends.sqlite3"
+    if using_sqlite:
+        use_prebuilt = click.confirm(
+            "It looks like you are using the default database backend (sqlite3). "
+            "Would you like to use a prebuilt-database with all third-party data "
+            "already loaded? If this is the first time you using the prebuild, "
+            "this will involve a ~3GB download and will unpack to roughly 7GB. A "
+            "backup of the download will be stored as well (so another ~3GB disk "
+            "space will be used). We recommend answering 'yes' for beginners."
+        )
+    else:
+        use_prebuilt = False
+
     # We can now proceed with reseting the database
     click.echo("Removing database and rebuilding...")
 
-    from simmate.configuration.django.database import reset_database
+    from simmate.database import connect
+    from simmate.database.utilities import reset_database
 
-    reset_database()
+    reset_database(use_prebuilt=use_prebuilt)
 
     # Let the user know everything succeeded
     click.echo("Success! Your database has been reset.")
@@ -49,7 +68,8 @@ def update():
     # check Django if there are any updates to be made
     click.echo("Checking for updates...")
 
-    from simmate.configuration.django.database import update_database
+    from simmate.database import connect
+    from simmate.database.utilities import update_database
 
     update_database()
 
@@ -70,7 +90,8 @@ def dumpdata(filename):
     # Begin writing the database to the json file.
     click.echo("Writing all data to JSON...")
 
-    from simmate.configuration.django.database import dump_database_to_json
+    from simmate.database import connect
+    from simmate.database.utilities import dump_database_to_json
 
     dump_database_to_json(filename=filename)
 
@@ -93,7 +114,8 @@ def loaddata(filename):
     # Begin writing the database to the json file.
     click.echo("Loading all data from JSON...")
 
-    from simmate.configuration.django.database import load_database_from_json
+    from simmate.database import connect
+    from simmate.database.utilities import load_database_from_json
 
     load_database_from_json(filename=filename)
 
@@ -103,5 +125,40 @@ def loaddata(filename):
     )
 
 
+@database.command()
+@click.option(
+    "--parallel",
+    is_flag=True,
+    help="(not allowed for SQLite3) loads data using parallel processing",
+)
+def load_remote_archives(parallel):
+    """
+    Goes through all third-party databases and loads their most recent remote
+    archives (if available). This utility helps with initializing a new
+    database build.
+
+    WARNING:
+    This can take several hours to run and there is no pause/continuation
+    implemented. This runs substantially faster when you are using a cloud
+    database backend (e.g. Postgres) and use the `--parallel` flag.
+
+    If you are using SQLite, we highly recommend using loading the default build
+    when you run `simmate database reset` instead.d.
+    """
+
+    from simmate.database import connect
+    from simmate.database.third_parties import load_remote_archives
+
+    load_remote_archives(parallel=parallel)
+    click.echo("Success! Your database now contains all third-party data.")
+
+
 # explicitly list functions so that pdoc doesn't skip them
-__all__ = ["database", "reset", "update", "dumpdata", "loaddata"]
+__all__ = [
+    "database",
+    "reset",
+    "update",
+    "dumpdata",
+    "loaddata",
+    "load_remote_archives",
+]
