+++ tutorials/Guides_for_contributors/environment.yaml
@@ -26,7 +26,9 @@ dependencies:
     - pandas >=1.3.5
     - pdoc >=11.0.0
     - plotly >=5.4.0
-    - prefect >=0.15.11
+    # - prefect >=2.0  # NOT AVAILABLE ON CONDA UNTIL STABLE RELEASE
+    - pip:
+        - prefect==2.0b7
     - pymatgen >=2022.1.9
     - pymatgen-analysis-diffusion >=2021.4.29
     - pytest >=6.2.5,<7.1
