+++ src/simmate/workflow_engine/workflow.py
@@ -5,23 +5,22 @@ import cloudpickle
 import yaml
 from typing import List
 
-# note: extra modules are imported from prefect for convenience imports elsewhere
-import prefect
-from prefect import Client
-from prefect import Flow as PrefectFlow, Parameter, task, Task
-from prefect.storage import Module as ModuleStorage
-from prefect.utilities.graphql import with_args
-from prefect.backend import FlowRunView, FlowView
+from prefect.tasks import task  # present only for convience imports elsewhere
+from prefect.flows import Flow
+from prefect.states import State
+from prefect.client import get_client
 
-# This import below is used to grab `watch_flow_run`, but we don't grab it directly
-# in order to support "mocking" this function in unit tests
-from prefect.backend import flow_run as flow_run_module
-
+import simmate
+from simmate.toolkit import Structure
 from simmate.database.base_data_types import Calculation
 from simmate.workflow_engine import S3Task
+from simmate.workflow_engine.common_tasks import (
+    load_input_and_register,
+    save_result,
+)
 
 
-class Workflow(PrefectFlow):
+class Workflow:
     """
     This class behaves exactly like a normal Prefect workflow, where we add some
     common utilities and pre-submit tasks. For example, there is the `run_cloud`
@@ -32,12 +31,20 @@ class Workflow(PrefectFlow):
     [prefect.core.flow.Flow](https://docs.prefect.io/api/latest/core/flow.html#flow-2)
     """
 
-    result_task: Task = None
+    # TODO: set storage attribute to module
+
+    # TODO: inherit doc from s3task
+    # by default we just copy the docstring of the S3task to the workflow
+    # workflow.__doc__ = s3task.__doc__
+
+    name: str = None
     """
-    In many ETL workflows, we may also want the result of the some terminating
-    task directly. This would save us from having to go find the result in
-    the database. So by setting `result_task`, we are able get access a specific task's
-    result -- directly as a python object(s). This is optional.
+    The name of this workflow (e.g. "relaxation.matproj")
+    """
+
+    version: str = simmate.__version__
+    """
+    Version number for this flow. Defaults to the Simmate version (such as "0.7.0")
     """
 
     project_name: str = None
@@ -76,13 +83,90 @@ class Workflow(PrefectFlow):
     A list of input parameters that should be used to register the calculation.
     """
 
+    @classmethod
+    def run(
+        cls,
+        structure: Structure,
+        command: str = None,
+        source: dict = None,
+        directory: str = None,
+        copy_previous_directory: bool = False,
+    ):
+        """
+        The workflow method, which can be overwritten when inheriting from this
+        class. This can be either a staticmethod or classmethod.
+
+        #### The default run method
+
+        There is a default run method implemented that is for S3Tasks.
+
+        Builds a workflow from a S3Task and it's corresponding database table.
+
+        Very often with Simmate's S3Tasks, the workflow for a single S3Task is
+        the same. The workflow is typically made of three tasks:
+
+        1. loading the input parameters and registering the calculation
+        2. running the calculation (what this S3Task does on its own)
+        3. saving the calculation results
+
+        Task 1 and 3 always use the same functions, where we just need to tell
+        it which database table we are registering/saving to.
+
+        Because of this common recipe for workflows, we use this method to make
+        the workflow for us.
+        """
+
+        parameters_cleaned = load_input_and_register(
+            structure=structure,
+            command=command,
+            source=source,
+            directory=directory,
+            copy_previous_directory=copy_previous_directory,
+        )
+
+        result = cls.s3task.run_as_prefect_task(
+            structure=parameters_cleaned["structure"],
+            command=parameters_cleaned["command"],
[…]
