+++ src/simmate/workflow_engine/supervised_staged_shell_task.py
@@ -6,7 +6,7 @@ import time
 import signal
 import subprocess
 import yaml
-from typing import List, Any
+from typing import List, Tuple
 
 import pandas
 
@@ -147,104 +147,93 @@ class S3Task:
     applied and none of the following handlers will be checked.
     """
 
-    def __init__(
-        self,
-        max_corrections: int = 5,
-        monitor: bool = True,
-        polling_timestep: float = 1,
-        monitor_freq: int = 300,
-        save_corrections_to_file: bool = True,
-        corrections_filename: str = "simmate_corrections.csv",
-        compress_output: bool = False,
-        **kwargs: Any,
-    ):
-        """
-        Creates a task instance of this class. The parameters passed will be the
-        same every time you call the task.run() method.
+    max_corrections: int = 5
+    """
+    The maximum number of times we can apply a correction and retry the shell
+    command. The maximum number of times that corrections will be made (and
+    shell command reran) before giving up on the calculation. Note, once this
+    limit is exceeded, the error is stored without correcting or restarting
+    the run.
+    """
 
-        Note, many of the parameters here are for the Monitoring settings. These
-        are only ever relevent if there are ErrorHandlers added that have
-        is_monitor=True. These handlers run while the shelltask itself is also
-        running. Read more about ErrorHandlers for more info.
+    monitor: bool = True
+    """
+    Whether to run monitor handlers while the command runs. False means
+    wait until the job has completed.
+    """
 
-        #### Parameters
+    polling_timestep: float = 1
+    """
+    If we are monitoring the job for errors while it runs, this is how often
+    (in seconds) we should check the status of our job. Note this check is
+    just whether the job is done or not. This is NOT how often we check for
+    errors. See monitor_freq for that.
+    """
 
-        - `max_corrections`:
-            The maximum number of times we can apply a correction and retry the shell
-            command. The maximum number of times that corrections will be made (and
-            shell command reran) before giving up on the calculation. Note, once this
-            limit is exceeded, the error is stored without correcting or restarting
-            the run.
-        - `monitor`:
-            Whether to run monitor handlers while the command runs. False means
-            wait until the job has completed.
-        - `polling_timestep`:
-            If we are monitoring the job for errors while it runs, this is how often
-            (in seconds) we should check the status of our job. Note this check is
-            just whether the job is done or not. This is NOT how often we check for
-            errors. See monitor_freq for that.
-        - `monitor_freq`:
-            The frequency we should run check for errors with our monitors. This is
-            based on the polling_timestep loops. For example, if we have a
-            polling_timestep of 10 seconds and a monitor_freq of 2, then we would run
-            the monitor checks every other loop -- or every 2x10 = 20 seconds. The
-            default values of polling_timestep=10 and monitor_freq=30 indicate that
-            we run monitoring functions every 5 minutes (10x30=300s=5min).
-        - `save_corrections_to_file`:
-            Whether to write a log file of the corrections made. The default is True.
-        - `corrections_filename`:
-            If save_corrections_to_file is True, this is the filename of where
-            to write the corrections. The default is "simmate_corrections.csv".
-        - `compress_output`:
-            Whether to compress the directory to a zip file at the end of the
-            task run. After compression, it will also delete the directory.
-            The default is False.
-        - `**kwargs`:
-            All extra arguments supported by
-            [prefect.core.task.Task](https://docs.prefect.io/api/latest/core/task.html).
-        """
+    monitor_freq: int = 300
+    """
+    The frequency we should run check for errors with our monitors. This is
+    based on the polling_timestep loops. For example, if we have a
+    polling_timestep of 10 seconds and a monitor_freq of 2, then we would run
+    the monitor checks every other loop -- or every 2x10 = 20 seconds. The
+    default values of polling_timestep=10 and monitor_freq=30 indicate that
[…]
