+++ src/simmate/toolkit/structure_prediction/evolution/database.py
@@ -4,8 +4,8 @@ from simmate.database.base_data_types import table_column, DatabaseTable
 
 from django.apps import apps as django_apps
 
-from prefect import Client
-from prefect.utilities.graphql import with_args
+# from prefect.client import Client
+# from prefect.utilities.graphql import with_args
 
 import plotly.graph_objects as plotly_go
 
