+++ src/simmate/calculators/vasp/workflows/static_energy/neb_endpoint.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from simmate.workflow_engine import s3task_to_workflow
+from simmate.workflow_engine import Workflow
 from simmate.calculators.vasp.tasks.static_energy import (
     NEBEndpointStaticEnergy as NEBEndpointStaticEnergyTask,
 )
@@ -8,12 +8,13 @@ from simmate.calculators.vasp.database.energy import (
     NEBEndpointStaticEnergy as NEBEndpointStaticEnergyResults,
 )
 
-workflow = s3task_to_workflow(
-    name="static-energy/neb-endpoint",
-    module=__name__,
-    project_name="Simmate-Energy",
-    s3task=NEBEndpointStaticEnergyTask,
-    database_table=NEBEndpointStaticEnergyResults,
-    register_kwargs=["structure", "source"],
-    description_doc_short="uses Materials Project settings meant for defect supercell structures",
-)
+
+class NEBEndpointStaticEnergy(Workflow):
+    name = "static-energy/neb-endpoint"
+    project_name = "Simmate-Energy"
+    s3task = NEBEndpointStaticEnergyTask
+    database_table = NEBEndpointStaticEnergyResults
+    register_kwargs = ["structure", "source"]
+    description_doc_short = (
+        "uses Materials Project settings meant for defect supercell structures"
+    )
