+++ src/simmate/calculators/vasp/workflows/electronic_structure/density_of_states.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+
+from simmate.workflow_engine import Workflow
+from simmate.calculators.vasp.tasks.density_of_states import (
+    MatProjDensityOfStates as MatProjDensityOfStatesTask,
+)
+from simmate.calculators.vasp.database.density_of_states import (
+    MatProjDensityOfStates as MatProjDensityOfStatesResults,
+)
+
+
+class MatProjDensityOfStates(Workflow):
+    name = "electronic-structure/density-of-states-matproj"
+    project_name = "Simmate-Electronic-Structure"
+    s3task = MatProjDensityOfStatesTask
+    database_table = MatProjDensityOfStatesResults
+    register_kwargs = ["structure", "source"]
+    description_doc_short = "uses Materials Project settings (PBE)"
