+++ src/simmate/calculators/vasp/workflows/relaxation/quality_01.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from simmate.workflow_engine import s3task_to_workflow
+from simmate.workflow_engine import Workflow
 from simmate.calculators.vasp.tasks.relaxation import (
     Quality01Relaxation as Quality01RelaxationTask,
 )
@@ -8,12 +8,13 @@ from simmate.calculators.vasp.database.relaxation import (
     Quality01Relaxation as Quality01RelaxationResults,
 )
 
-workflow = s3task_to_workflow(
-    name="relaxation/quality01",
-    module=__name__,
-    project_name="Simmate-Relaxation",
-    s3task=Quality01RelaxationTask,
-    database_table=Quality01RelaxationResults,
-    register_kwargs=["structure", "source"],
-    description_doc_short="low-quality settings meant for highly unreasonable structures",
-)
+
+class Quality01Relaxation(Workflow):
+    name = "relaxation/quality01"
+    project_name = "Simmate-Relaxation"
+    s3task = Quality01RelaxationTask
+    database_table = Quality01RelaxationResults
+    register_kwargs = ["structure", "source"]
+    description_doc_short = (
+        "low-quality settings meant for highly unreasonable structures"
+    )
