### 2022-07-06 8026534
add required_files tag
