+++ src/simmate/calculators/vasp/workflows/relaxation/staged.py
@@ -2,149 +2,35 @@
 
 import os
 
-from simmate.workflow_engine.workflow import (
-    Workflow,
-    Parameter,
-    ModuleStorage,
-)
+from simmate.workflow_engine import Workflow
 from simmate.workflow_engine.common_tasks import load_input_and_register
-from simmate.calculators.vasp.workflows.relaxation.quality_00 import (
-    workflow as relaxation_quality00,
+from simmate.calculators.vasp.workflows.relaxation.quality00 import (
+    Relaxation__VASP__Quality00,
 )
-from simmate.calculators.vasp.workflows.relaxation.quality_01 import (
-    workflow as relaxation_quality01,
+from simmate.calculators.vasp.workflows.relaxation.quality01 import (
+    Relaxation__VASP__Quality01,
 )
-from simmate.calculators.vasp.workflows.relaxation.quality_02 import (
-    workflow as relaxation_quality02,
+from simmate.calculators.vasp.workflows.relaxation.quality02 import (
+    Relaxation__VASP__Quality02,
 )
-from simmate.calculators.vasp.workflows.relaxation.quality_03 import (
-    workflow as relaxation_quality03,
+from simmate.calculators.vasp.workflows.relaxation.quality03 import (
+    Relaxation__VASP__Quality03,
 )
-from simmate.calculators.vasp.workflows.relaxation.quality_04 import (
-    workflow as relaxation_quality04,
+from simmate.calculators.vasp.workflows.relaxation.quality04 import (
+    Relaxation__VASP__Quality04,
 )
-from simmate.calculators.vasp.workflows.static_energy.quality_04 import (
-    workflow as energy_quality04,
+from simmate.calculators.vasp.workflows.static_energy.quality04 import (
+    Static_Energy__VASP__Quality04,
 )
 
 from simmate.calculators.vasp.database.relaxation import StagedRelaxation
-from simmate.calculators.vasp.database.energy import Quality04StaticEnergy
 
 
-# init workflow tasks
-# OPTIMIZE: Make this a for-loop in Prefect 2.0! We can use a for-loop in the
-# workflow context below too.
-relax_task_00 = relaxation_quality00.to_workflow_task()
-relax_task_01 = relaxation_quality01.to_workflow_task()
-relax_task_02 = relaxation_quality02.to_workflow_task()
-relax_task_03 = relaxation_quality03.to_workflow_task()
-relax_task_04 = relaxation_quality04.to_workflow_task()
-static_task_04 = energy_quality04.to_workflow_task()
-
-with Workflow("relaxation/staged") as workflow:
-
-    structure = Parameter("structure")
-    command = Parameter("command", default="vasp_std > vasp.out")
-    source = Parameter("source", default=None)
-    directory = Parameter("directory", default=None)
-    copy_previous_directory = Parameter(
-        "copy_previous_directory",
-        default=False,
-    )
-
-    parameters_cleaned = load_input_and_register(
-        structure=structure,
-        command=command,
-        source=source,
-        directory=directory,
-        copy_previous_directory=copy_previous_directory,
-    )
-
-    # Our first relaxation is directly from our inputs. The remaining one
-    # pass along results
-    run_id_00 = relax_task_00(
-        structure=parameters_cleaned["structure"],
-        command=parameters_cleaned["command"],
-        directory=parameters_cleaned["directory"] + os.path.sep + "relax_00",
-    )
-
-    # TODO: Use a for-loop in Prefect 2.0!
-
-    # relaxation 01
-    run_id_01 = relax_task_01(
-        structure={
-            "database_table": "Quality00Relaxation",
-            "directory": run_id_00["directory"],
-            "structure_field": "structure_final",
-        },
[…]
