+++ src/simmate/calculators/vasp/workflows/relaxation/__init__.py
@@ -1,11 +1,1 @@
 # -*- coding: utf-8 -*-
-
-from .mit import workflow as mit_workflow
-from .materials_project import workflow as matproj_workflow
-from .quality_00 import workflow as quality00_workflow
-from .quality_01 import workflow as quality01_workflow
-from .quality_02 import workflow as quality02_workflow
-from .quality_03 import workflow as quality03_workflow
-from .quality_04 import workflow as quality04_workflow
-from .staged import workflow as staged_workflow
-from .neb_endpoint import workflow as neb_endpoint_workflow
