+++ src/simmate/workflow_engine/workflow.py
@@ -8,8 +8,8 @@ from typing import List
 from prefect.tasks import task  # present only for convience imports elsewhere
 from prefect.flows import Flow
 from prefect.states import State
-
-# from prefect.client import get_client
+from prefect.context import FlowRunContext
+from prefect.client import get_client
 
 import simmate
 from simmate.toolkit import Structure
@@ -77,7 +77,7 @@ class Workflow:
     """
 
     @classmethod
-    def run(
+    def run_config(
         cls,
         structure: Structure,
         command: str = None,
@@ -144,13 +144,13 @@ class Workflow:
         Converts this workflow into a Prefect flow
         """
         return Flow(
-            fn=cls.run,
+            fn=cls.run_config,
             name=cls.name_full,
             version=cls.version,
         )
 
     @classmethod
-    def run_as_prefect_flow(cls, **kwargs) -> State:
+    def run(cls, **kwargs) -> State:
         """
         A convience method to run a workflow as a subflow in a prefect context.
         """
@@ -235,119 +235,8 @@ class Workflow:
         # use yaml to make the printout pretty (no quotes and separate lines)
         print(yaml.dump(cls.parameter_names))
 
-    def run_cloud(
-        self,
-        labels: List[str] = [],
-        wait_for_run: bool = True,
-        **kwargs,
-    ) -> str:
-        """
-        This schedules the workflow to run remotely on Prefect Cloud.
-
-        #### Parameters
-
-        - `labels`:
-            a list of labels to schedule the workflow with
-
-        - `wait_for_run`:
-            whether to wait for the workflow to finish. If False, the workflow
-            will simply be submitted and then exit. The default is True.
-
-        - `**kwargs`:
-            all options that are normally passed to the workflow.run() method
-
-        #### Returns
-
-        - The flow run id that was used in prefect cloud.
-
-
-        #### Usage
-
-        Make sure you have Prefect properly configured and have registered your
-        workflow with the backend.
-
-        Note that this method can be viewed as a fork of:
-            - from prefect.tasks.prefect.flow_run import create_flow_run
-        It can also be viewed as a more convenient way to call to client.create_flow_run.
-        I don't accept any other client.create_flow_run() inputs besides 'labels'.
-        This may change in the future if I need to set flow run names or schedules.
-        """
-        raise NotImplementedError("Migrating to Prefect 2.0")
-
-        # Grab the logger as we will print useful information below.
-        # In some cases, there is no logger present, so we just set logger=None.
-        # This possibility is why we use `if logger else None` statements below.
-        logger = getattr(prefect.context, "logger", None)
-
-        # Grab the Flow's ID from Prefect Cloud
-        logger.debug("Looking up flow metadata...") if logger else None
-        flow_view = FlowView.from_flow_name(
-            self.name,
-            project_name=self.project_name,
-        )
-
-        # Prefect does not properly deserialize objects that have
-        # as as_dict or to_dict method, so we use a custom method to do that here
-        parameters_serialized = self._serialize_parameters(**kwargs)
-        # BUG: What if we are submitting using a filename? We don't want to
-        # submit to a cluster and have the job fail because it doesn't have
-        # access to the file. One solution could be to deserialize right before
-        # serializing in the next line in order to ensure parameters that
-        # accept file names are submitted with all necessary data.
-
-        # Now we submit the workflow
-        logger.info(f"Creating flow run for {self.name}...") if logger else None
-        client = Client()
-        flow_run_id = client.create_flow_run(
-            flow_id=flow_view.flow_id,
-            parameters=parameters_serialized,
-            labels=labels,
-        )
-
-        # we log the website url to the flow for the user
-        run_url = client.get_cloud_url("flow-run", flow_run_id)
-        logger.info(f"Created flow run: {run_url}") if logger else None
-
-        # -----------------------------------
[…]
