+++ src/simmate/calculators/vasp/workflows/relaxation/all.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 
-from .matproj import Relaxation__VASP__MatProj
-from .mit import Relaxation__VASP__MIT
-from .neb_endpoint import Relaxation__VASP__NEB_Endpoint
-from .quality00 import Relaxation__VASP__Quality00
-from .quality01 import Relaxation__VASP__Quality01
-from .quality02 import Relaxation__VASP__Quality02
-from .quality03 import Relaxation__VASP__Quality03
-from .quality04 import Relaxation__VASP__Quality04
+from .matproj import Relaxation__Vasp__MatProj
+from .mit import Relaxation__Vasp__Mit
+from .neb_endpoint import Relaxation__Vasp__NebEndpoint
+from .quality00 import Relaxation__Vasp__Quality00
+from .quality01 import Relaxation__Vasp__Quality01
+from .quality02 import Relaxation__Vasp__Quality02
+from .quality03 import Relaxation__Vasp__Quality03
+from .quality04 import Relaxation__Vasp__Quality04
