+++ src/simmate/calculators/vasp/workflows/relaxation/quality00.py
@@ -9,7 +9,7 @@ from simmate.calculators.vasp.database.relaxation import (
 )
 
 
-class Relaxation__VASP__Quality00(Workflow):
+class Relaxation__Vasp__Quality00(Workflow):
     s3task = Quality00RelaxationTask
     database_table = Quality00RelaxationResults
     register_kwargs = ["structure", "source"]
