+++ src/simmate/calculators/vasp/workflows/all.py
@@ -1,19 +1,19 @@
 # -*- coding: utf-8 -*-
 
 from .static_energy.all import (
-    Static_Energy__VASP__MatProj,
-    Static_Energy__VASP__MIT,
-    Static_Energy__VASP__Quality04,
-    Static_Energy__VASP__NEB_Endpoint,
+    StaticEnergy__Vasp__MatProj,
+    StaticEnergy__Vasp__Mit,
+    StaticEnergy__Vasp__Quality04,
+    StaticEnergy__Vasp__NebEndpoint,
 )
 
 from .relaxation.all import (
-    Relaxation__VASP__MatProj,
-    Relaxation__VASP__MIT,
-    Relaxation__VASP__NEB_Endpoint,
-    Relaxation__VASP__Quality00,
-    Relaxation__VASP__Quality01,
-    Relaxation__VASP__Quality02,
-    Relaxation__VASP__Quality03,
-    Relaxation__VASP__Quality04,
+    Relaxation__Vasp__MatProj,
+    Relaxation__Vasp__Mit,
+    Relaxation__Vasp__NebEndpoint,
+    Relaxation__Vasp__Quality00,
+    Relaxation__Vasp__Quality01,
+    Relaxation__Vasp__Quality02,
+    Relaxation__Vasp__Quality03,
+    Relaxation__Vasp__Quality04,
 )
