+++ src/simmate/calculators/vasp/workflows/electronic_structure/all.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from .matproj_full import ElectronicStructure__Vasp__MatProjFull
+
+# NOTE: we do not import the independent band structure and density of states
+# workflows as these should not be called directly.
