+++ src/simmate/workflows/electronic_structure.py
@@ -8,6 +8,6 @@ for a given crystal structure
 # TODO: when I add more calculators, I can do something like this...
 # if "simmate.calculators.vasp" in installed_apps:
 
-from simmate.calculators.vasp.workflows.electronic_structure import (
-    matproj_workflow,
+from simmate.calculators.vasp.workflows.electronic_structure.all import (
+    ElectronicStructure__Vasp__MatProjFull,
 )
