### 2022-07-07 ba0a2c6
update dynamics and changelog
