+++ src/simmate/calculators/vasp/workflows/population_analysis/__init__.py
@@ -1,5 +1,1 @@
 # -*- coding: utf-8 -*-
-
-from .bader import prebader_matproj_workflow, workflow as bader_matproj_workflow
-from .elf import workflow as elf_matproj_workflow
-from .badelf import prebadelf_matproj_workflow, workflow as badelf_matproj_workflow
