+++ src/simmate/calculators/vasp/workflows/electronic_structure/__init__.py
@@ -1,3 +1,1 @@
 # -*- coding: utf-8 -*-
-
-from .materials_project import workflow as matproj_workflow
