+++ src/simmate/calculators/vasp/workflows/all.py
@@ -1,5 +1,22 @@
 # -*- coding: utf-8 -*-
 
+from .diffusion.all import (
+    Diffusion__Vasp__NebAllPaths,
+    Diffusion__Vasp__NebSinglePath,
+    Diffusion__Vasp__NebFromEndpoints,
+    Diffusion__Vasp__NebFromImages,
+)
+
+from .electronic_structure.all import (
+    ElectronicStructure__Vasp__MatProjFull,
+)
+
+from .population_analysis.all import (
+    PopulationAnalysis__Vasp__BaderMatproj,
+    PopulationAnalysis__Vasp__ElfMatproj,
+    PopulationAnalysis__Vasp__BadelfMatproj,
+)
+
 from .static_energy.all import (
     StaticEnergy__Vasp__MatProj,
     StaticEnergy__Vasp__Mit,
@@ -16,15 +33,9 @@ from .relaxation.all import (
     Relaxation__Vasp__Quality02,
     Relaxation__Vasp__Quality03,
     Relaxation__Vasp__Quality04,
+    Relaxation__Vasp__Staged,
 )
 
-from .electronic_structure.all import (
-    ElectronicStructure__Vasp__MatProjFull,
-)
+from .customized import Customized__Vasp__UserConfig
 
-from .diffusion.all import (
-    Diffusion__Vasp__NebAllPaths,
-    Diffusion__Vasp__NebSinglePath,
-    Diffusion__Vasp__NebFromEndpoints,
-    Diffusion__Vasp__NebFromImages,
-)
+from .dynamics import Dynamics__Vasp__Mit
