+++ src/simmate/calculators/vasp/workflows/_to_do/scratch_pdos_empty.py
@@ -1,549 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import subprocess
-from pymatgen.io.vasp.sets import DictSet
-
-#!!! The warning that is raised here is because there is no YAML! It can be ignored
-class StaticEnergyCalc(DictSet):
-
-    CONFIG = {
-        "INCAR": {
-            "EDIFF": 1.0e-07,
-            "EDIFFG": -1e-04,
-            "ENCUT": 520,
-            "ISIF": 3,  # !!! do I want this..?
-            "ISMEAR": 0,  # Guassian smearing #!!! read docs!
-            "LCHARG": True,  # write CHGCAR
-            "LAECHG": True,  # write AECCAR0, AECCAR1, and AECCAR2
-            "LWAVE": True,  # write WAVECAR
-            "NSW": 0,  # single energy calc
-            "PREC": "Accurate",  # !!! USE Accurate WHEN NOT DOING BADELF
-            "IVDW": 12,  # van der waals correction
-            "ISMEAR": 0,  # Guassian smearing = 0 if system unknown!!!!!!!!!!!
-            "SIGMA": 0.060,
-            # 'NBANDS': 643, # Calculate more bands than normal (extra empty)
-            "SYMPREC": 1e-8,  # !!! CUSTODIAN FIX - dont use unless needed
-            # 'ISYM': 0,
-            # 'NGX': 100,
-            # 'NGY': 100,
-            # 'NGZ': 100,
-            "NGXF": 100,
-            "NGYF": 100,
-            "NGZF": 100,
-            "NCORE": 8,
-            #!!! TESTING
-            # ELFCAR (optional)
-            # 'LELF': True, # write ELFCAR
-            # 'NPAR': 1, # Must be set if LELF is set to True
-        },
-        "KPOINTS": {"reciprocal_density": 100},
-        "POTCAR_FUNCTIONAL": "PBE_54",
-        "POTCAR": {
-            "Ac": "Ac",
-            "Ag": "Ag",
-            "Al": "Al",
-            "Ar": "Ar",
-            "As": "As",
-            "Au": "Au",
-            "B": "B",
-            "Ba": "Ba_sv",
-            "Be": "Be_sv",
-            "Bi": "Bi",
-            "Br": "Br",
-            "C": "C",
-            "Ca": "Ca_sv",
-            "Cd": "Cd",
-            "Ce": "Ce",
-            "Cl": "Cl",
-            "Co": "Co",
-            "Cr": "Cr_pv",
-            "Cs": "Cs_sv",
-            "Cu": "Cu_pv",
-            "Dy": "Dy_3",
-            "Er": "Er_3",
-            "Eu": "Eu",
-            "F": "F",
-            "Fe": "Fe_pv",
-            "Ga": "Ga_d",
-            "Gd": "Gd",
-            "Ge": "Ge_d",
-            "H": "H",
-            "He": "He",
-            "Hf": "Hf_pv",
-            "Hg": "Hg",
-            "Ho": "Ho_3",
-            "I": "I",
-            "In": "In_d",
-            "Ir": "Ir",
-            "K": "K_sv",
-            "Kr": "Kr",
-            "La": "La",
-            "Li": "Li_sv",
-            "Lu": "Lu_3",
-            "Mg": "Mg_pv",
-            "Mn": "Mn_pv",
-            "Mo": "Mo_pv",
-            "N": "N",
-            "Na": "Na_pv",
-            "Nb": "Nb_pv",
-            "Nd": "Nd_3",
-            "Ne": "Ne",
-            "Ni": "Ni_pv",
-            "Np": "Np",
-            "O": "O",
[…]
