+++ src/simmate/calculators/vasp/workflows/electronic_structure/matproj_full.py
@@ -5,27 +5,27 @@ import os
 from simmate.workflow_engine import Workflow
 from simmate.workflow_engine.common_tasks import load_input_and_register
 
-from simmate.calculators.vasp.workflows.static_energy.matproj import (
-    StaticEnergy__Vasp__MatProj as static_workflow,
+from simmate.calculators.vasp.workflows.static_energy.Matproj import (
+    StaticEnergy__Vasp__Matproj as static_workflow,
 )
-from simmate.calculators.vasp.workflows.electronic_structure.matproj_density_of_states import (
-    ElectronicStructure__Vasp__MatProjDensityOfStates as dos_workflow,
+from simmate.calculators.vasp.workflows.electronic_structure.Matproj_density_of_states import (
+    ElectronicStructure__Vasp__MatprojDensityOfStates as dos_workflow,
 )
-from simmate.calculators.vasp.workflows.electronic_structure.matproj_band_structure import (
-    ElectronicStructure__Vasp__MatProjBandStructure as bandstruct_workflow,
+from simmate.calculators.vasp.workflows.electronic_structure.Matproj_band_structure import (
+    ElectronicStructure__Vasp__MatprojBandStructure as bandstruct_workflow,
 )
 
 
-class ElectronicStructure__Vasp__MatProjFull(Workflow):
+class ElectronicStructure__Vasp__MatprojFull(Workflow):
     """
     Runs a static energy calculation followed by non-SCF calculations for
     band structure and density of states.
 
     This is therefore a "Nested Workflow" made of the following smaller workflows:
 
-        - static-energy.vasp.matproj
-        - band-structure.vasp.matproj
-        - density-of-states.vasp.matproj
+        - static-energy.vasp.Matproj
+        - band-structure.vasp.Matproj
+        - density-of-states.vasp.Matproj
 
     Note, these calculations are done using PBE, which is known to underestimate
     band gaps. For higher quality electronic structure calculations, you should
