+++ src/simmate/calculators/vasp/tasks/test/test_band_structure.py
@@ -4,14 +4,14 @@ import os
 
 from simmate.conftest import copy_test_files, make_dummy_files
 from simmate.calculators.vasp.inputs import Potcar
-from simmate.calculators.vasp.tasks.band_structure import MatProjBandStructure
+from simmate.calculators.vasp.tasks.band_structure import MatprojBandStructure
 from simmate.calculators.vasp.inputs.potcar_mappings import PBE_ELEMENT_MAPPINGS
 
 
 def test_band_structure_setup(structure, tmpdir, mocker):
 
     # init with default settings
-    task = MatProjBandStructure()
+    task = MatprojBandStructure()
 
     # estabilish filenames that we make and commonly reference
     incar_filename = os.path.join(tmpdir, "INCAR")
@@ -47,7 +47,7 @@ def test_band_structure_workup(tmpdir):
     )
 
     # init main task
-    task = MatProjBandStructure()
+    task = MatprojBandStructure()
 
     # estabilish filenames that we make and commonly reference
     summary_filename = os.path.join(tmpdir, "simmate_summary.yaml")
