+++ src/simmate/calculators/vasp/tasks/density_of_states/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 
 from .base import VaspDensityOfStates
-from .materials_project import MatProjDensityOfStates
-from .materials_project_hse import MatProjHSEDensityOfStates
+from .materials_project import MatprojDensityOfStates
+from .materials_project_hse import MatprojHSEDensityOfStates
