+++ src/simmate/calculators/vasp/workflows/electronic_structure/matproj_density_of_states.py
@@ -2,14 +2,14 @@
 
 from simmate.workflow_engine import Workflow
 from simmate.calculators.vasp.tasks.density_of_states import (
-    MatProjDensityOfStates as MatProjDensityOfStatesTask,
+    MatprojDensityOfStates as MatprojDensityOfStatesTask,
 )
 from simmate.calculators.vasp.database.density_of_states import (
-    MatProjDensityOfStates as MatProjDensityOfStatesResults,
+    MatprojDensityOfStates as MatprojDensityOfStatesResults,
 )
 
 
-class ElectronicStructure__Vasp__MatProjDensityOfStates(Workflow):
-    s3task = MatProjDensityOfStatesTask
-    database_table = MatProjDensityOfStatesResults
+class ElectronicStructure__Vasp__MatprojDensityOfStates(Workflow):
+    s3task = MatprojDensityOfStatesTask
+    database_table = MatprojDensityOfStatesResults
     description_doc_short = "uses Materials Project settings (PBE)"
