+++ src/simmate/workflows/test/test_workflows_utilities.py
@@ -18,10 +18,10 @@ def test_get_workflow_types():
         "customized",
         "diffusion",
         "dynamics",
-        "electronic_structure",
-        "population_analysis",
+        "electronic-structure",
+        "population-analysis",
         "relaxation",
-        "static_energy",
+        "static-energy",
     ]
 
 
@@ -34,11 +34,11 @@ def test_list_of_all_workflows():
         "diffusion.vasp.neb-from-images",
         "diffusion.vasp.neb-single-path",
         "dynamics.vasp.mit",
-        "electronic-structure.vasp.mat-proj-full",
+        "electronic-structure.vasp.matproj-full",
         "population-analysis.vasp.badelf-matproj",
         "population-analysis.vasp.bader-matproj",
         "population-analysis.vasp.elf-matproj",
-        "relaxation.vasp.mat-proj",
+        "relaxation.vasp.matproj",
         "relaxation.vasp.mit",
         "relaxation.vasp.neb-endpoint",
         "relaxation.vasp.quality00",
@@ -46,7 +46,7 @@ def test_list_of_all_workflows():
         "relaxation.vasp.quality02",
         "relaxation.vasp.quality03",
         "relaxation.vasp.quality04",
-        "static-energy.vasp.mat-proj",
+        "static-energy.vasp.matproj",
         "static-energy.vasp.mit",
         "static-energy.vasp.neb-endpoint",
         "static-energy.vasp.quality04",
@@ -68,9 +68,11 @@ def test_list_of_workflows_by_type():
 
 def test_get_workflow():
 
-    from simmate.workflows.static_energy import mit_workflow
+    from simmate.workflows.static_energy import (
+        StaticEnergy__Vasp__Matproj as workflow,
+    )
 
-    assert get_workflow("static-energy/mit") == mit_workflow
+    assert get_workflow("static-energy.vasp.mit") == workflow
 
 
 def test_get_unique_paramters():
