+++ src/simmate/calculators/vasp/database/relaxation.py
@@ -29,7 +29,7 @@ MITRelaxation, MITIonicStep = Relaxation.create_subclasses("MIT", module=__name_
 (
     MatVirtualLabCINEBEndpointRelaxation,
     MatVirtualLabCINEBEndpointIonicStep,
-) = Relaxation.create_subclasses("NEBEndpoint", module=__name__)
+) = Relaxation.create_subclasses("MatVirtualLabCINEBEndpoint", module=__name__)
 
 Quality00Relaxation, Quality00IonicStep = Relaxation.create_subclasses(
     "Quality00",
