+++ src/simmate/calculators/vasp/tasks/nudged_elastic_band/base.py
@@ -31,14 +31,6 @@ class VaspNudgedElasticBandTask(VaspTask):
     may be useful if you'd like to make your own variation of this class.
     """
 
-    requires_structure = False
-    """
-    `structure` is not a required parameter for the `run` method. This is a 
-    unique case for VASP calculations because the input is NOT a single 
-    structure, but instead a list of structures -- spefically a list
-    supercell images along the diffusion pathway.
-    """
-
     # Pymatgen's NEB parser does not read from the vasprun.xml so it can't
     # confirm convergence here. I'll have to write my own output class to do this.
     confirm_convergence = False
@@ -46,9 +38,8 @@ class VaspNudgedElasticBandTask(VaspTask):
     @classmethod
     def _pre_checks(
         cls,
-        structures: MigrationImages,
+        migration_images: MigrationImages,
         directory: str,
-        structure: None,
     ):
         """
         Runs a series of checks to ensure the user configured the job correctly.
@@ -56,14 +47,6 @@ class VaspNudgedElasticBandTask(VaspTask):
         This is called automatically within the setup() method and shouldn't be
         used directly.
         """
-
-        # The first common mistake is providing a structure instead of structures
-        if structure:
-            raise Exception(
-                "This NEB calculation requires a list of structures (aka images). "
-                "Make sure you provide the structures input and NOT a single "
-                "structure."
-            )
 
         # The next common mistake is to mislabel the number of images in the
         # INCAR file.
@@ -73,7 +56,7 @@ class VaspNudgedElasticBandTask(VaspTask):
             # if so, we check that it was set correctly. It should be equal to
             # the number of structures minus 2 (because we don't count the
             # start and end images here.)
-            if nimages != (len(structures) - 2):
+            if nimages != (len(migration_images) - 2):
                 raise Exception(
                     "IMAGES looks to be improperly set! This value should not"
                     " include the start/end images -- so make sure you counted"
@@ -90,12 +73,11 @@ class VaspNudgedElasticBandTask(VaspTask):
         #       "M_divide: can not subdivide 16 nodes by 3"
 
         # make sure all images are contained with the cell
-        cls.structures = cls._process_structures(structures)
+        migration_images_cleaned = cls._process_structures(migration_images)
+        return migration_images_cleaned
 
     @staticmethod
-    def _process_structures(
-        structures: MigrationImages,
-    ):
+    def _process_structures(structures: MigrationImages):
         """
         Remove any atom jumps across the cell.
 
@@ -113,14 +95,14 @@ class VaspNudgedElasticBandTask(VaspTask):
                 if numpy.any(numpy.abs(t) > 0.5):
                     s.translate_sites([i], t, to_unit_cell=False)
             structures.append(s)
-        return structures
+        return MigrationImages(structures) # convert back to simmate object
 
     @classmethod
     def setup(
         cls,
-        structure: None,  # This is first and required bc of S3task.run
         directory: str,
-        structures: MigrationImages,
+        migration_images: MigrationImages,
+        **kwargs,
     ):
         """
         Writes input files for a NEB calculation. Each structure image recieves
@@ -148,12 +130,12 @@ class VaspNudgedElasticBandTask(VaspTask):
         # removing it from the S3Task...
 
         # run some prechecks to make sure the user has everything set up properly.
-        cls._pre_checks(structures, directory, structure)
+        migration_images_cleaned = cls._pre_checks(migration_images, directory)
 
         # Here, each image (start to end structures) is put inside of its own
         # folder. We make those folders here, where they are named 00, 01, 02...N
[…]
