+++ src/simmate/workflow_engine/supervised_staged_shell_task.py
@@ -7,7 +7,7 @@ import signal
 import subprocess
 import yaml
 from typing import List, Tuple
-from functool import cache
+from functools import cache
 
 import pandas
 
