+++ src/simmate/command_line/workflows.py
@@ -110,15 +110,15 @@ def explore():
 
     click.echo("\nGathering all available workflows...")
     from simmate.workflows.utilities import (
-        WORKFLOW_TYPES,
+        get_workflow_types,
         get_list_of_workflows_by_type,
         get_workflow,
     )
 
     click.echo("\n\nWhat type of analysis are you interested in?")
-    types_cleaned = [t.replace("-", " ") for t in WORKFLOW_TYPES]
-    type_index = list_options(types_cleaned)
-    selected_type = WORKFLOW_TYPES[type_index]
+    workflow_types = get_workflow_types()
+    type_index = list_options(workflow_types)
+    selected_type = workflow_types[type_index]
 
     # TODO: have the user select a calculator for this analysis. For now,
     # we are assuming VASP because those are our only workflows
@@ -128,7 +128,10 @@ def explore():
     present_index = list_options(presets)
     selected_preset = presets[present_index]
 
-    final_workflow_name = selected_type + "/" + selected_preset
+    final_workflow_name = selected_type + ".vasp." + selected_preset
+    # BUG: I assume vasp for now, but this should change when I add workflows
+    # from new calculators.
+
     click.echo(f"\n\n===================== {final_workflow_name} =====================")
 
     # now we load this workflow and print the docstring.
@@ -138,6 +141,7 @@ def explore():
         print_equivalent_import=True,
     )
 
+    click.echo("Description:")
     click.echo(workflow.__doc__)
 
     click.echo("Parameters:")
@@ -218,7 +222,8 @@ def setup_only(context, workflow_name):
         precheck_flow_exists=True,
         print_equivalent_import=True,
     )
-    kwargs_cleaned = parse_parameters(context=context)
+    kwargs_input = parse_parameters(context=context)
+    kwargs_cleaned = workflow._deserialize_parameters(**kwargs_input)
 
     click.echo("WRITING INPUT FILES...")
 
@@ -272,7 +277,7 @@ def run(context, workflow_name):
     result = workflow.run(**kwargs_cleaned)
 
     # Let the user know everything succeeded
-    if result.is_successful():
+    if result.is_completed():
         click.echo("Success! All results are also stored in your database.")
 
 
@@ -303,7 +308,7 @@ def run_cloud(context, workflow_name):
     result = workflow.run_cloud(**kwargs_cleaned)
 
     # Let the user know everything succeeded
-    if result.is_successful():
+    if result.is_completed():
         click.echo("Success! All results are also stored in your database.")
 
 
@@ -351,7 +356,11 @@ def run_yaml(filename):
 
         for key, update_values in list(parameters.items()):
             # Skip the base keys
-            if key in ["workflow_base", "input_parameters", "updated_settings"]:
+            if key in [
+                "workflow_base",
+                "input_parameters",
+                "updated_settings",
+            ]:
                 continue
             # if there is no prefix, then we have a normal input parameter
             elif not key.startswith("custom__"):
@@ -368,7 +377,7 @@ def run_yaml(filename):
     result = workflow.run(**parameters)
 
     # Let the user know everything succeeded
-    if result.is_successful():
+    if result.is_completed():
         click.echo("Success! All results are also stored in your database.")
 
 
