+++ src/simmate/calculators/vasp/tasks/test/test_nudged_elastic_band.py
@@ -20,9 +20,6 @@ def test_neb_setup(sample_structures, tmpdir, mocker):
     structure = sample_structures["Y2CI2_mp-1206803_primitive"]
     images = MigrationImages.from_structure(structure, "I")[0]
 
-    # init with default settings
-    task = MITNudgedElasticBand()
-
     # estabilish filenames that we make and commonly reference
     incar_filename = os.path.join(tmpdir, "INCAR")
     potcar_filename = os.path.join(tmpdir, "POTCAR")
@@ -40,9 +37,8 @@ def test_neb_setup(sample_structures, tmpdir, mocker):
     )
 
     # try to make input files in the tmpdir
-    task.setup(
-        structure=None,
-        structures=images,
+    MITNudgedElasticBand.setup(
+        migration_images=images,
         directory=tmpdir,
     )
     assert os.path.exists(incar_filename)
