+++ tutorials/Guides_for_contributors/environment.yaml
@@ -22,6 +22,7 @@ dependencies:
     - django-filter >=21.1
     - djangorestframework >=3.13.1
     - matminer >=0.7.6
+    - nest-asyncio >1.5.5
     - numpy >=1.22.0
     - pandas >=1.3.5
     - pdoc >=11.0.0
