### 2022-07-11 93e71a7
begin debug series
