### 2022-07-11 3c18f8d
debug with result calls
