### 2022-07-11 10b7f00
return to parallel runs and disable mac
