+++ .github/workflows/lint-and-test.yml
@@ -30,7 +30,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ["ubuntu-latest", "macos-latest", "windows-latest"]
+        os: ["ubuntu-latest", "windows-latest"]  # "macos-latest", BUG
         python-version: ["3.10"]
 
     steps:
