+++ src/simmate/calculators/vasp/workflows/diffusion/test/test_all_paths.py
@@ -41,7 +41,7 @@ def test_neb(sample_structures, tmpdir, mocker):
 
     # BUG: If this job hangs, this line can break that can have it exit
     # the pytest command successfully. Still trying to figure out why.
-    #   Diffusion__Vasp__NebAllPaths.nflows_submitted
+    Diffusion__Vasp__NebAllPaths.nflows_submitted
     #
     # As a minimal example, I can get a test to hang in a lone test using...
     #
