+++ src/simmate/calculators/vasp/tasks/nudged_elastic_band/base.py
@@ -264,7 +264,11 @@ class VaspNudgedElasticBandTask(VaspTask):
         # plot the results
         # plot = neb_results.get_plot()
         # plot.savefig(os.path.join(directory, "NEB_plot.jpeg"))
-        print("NEB plot generation is temporarily disabled due to a bug in pymatgen")
+        print(
+            "NEB plot generation is temporarily disabled due to a bug in prefect "
+            "where matplotlib.plot() leads to a segmentation fault. "
+            "See https://github.com/PrefectHQ/prefect/issues/5991"
+        )
 
         # convert all the structures to a MigrationImages object so we can write
         # the summed structure.
