+++ src/simmate/calculators/bader/tasks/badelf.py
@@ -17,7 +17,7 @@ class BaderELFAnalysis(BaderAnalysis):
 
     command = "bader CHGCAR_empty -ref ELFCAR_empty > bader.out"
     requires_structure = True  # !!! This may change in the future
-    required_files = BaderAnalysis.required_files + ["ELFCAR"]
+    required_files = ["CHGCAR", "ELFCAR"]
 
     @staticmethod
     def setup(structure: Structure, directory: str):
@@ -32,9 +32,6 @@ class BaderELFAnalysis(BaderAnalysis):
         - `structure`:
             The structure to use when rewriting the ELFCAR and CHGCAR files.
         """
-
-        # Run the default method of the parent class.
-        super().setup(structure=structure, directory=directory)
 
         # establish file names
         elfcar_filename = os.path.join(directory, "ELFCAR")
