+++ src/simmate/database/base_data_types/test/test_relaxation_db.py
@@ -19,14 +19,14 @@ def test_static_energy_table(structure):
 
     # test writing to database
     structure_db = TestRelaxation.from_prefect_id(
-        id="example-id-123",
+        prefect_flow_run_id="example-id-123",
         structure=structure,
     )
     structure_db.save()
 
     # try grabbing the calculation again and make sure it loaded from the
     # database rather than creating a new entry
-    structure_db2 = TestRelaxation.from_prefect_id(id="example-id-123")
+    structure_db2 = TestRelaxation.from_prefect_id(prefect_flow_run_id="example-id-123")
     assert structure_db.id == structure_db2.id
 
     # test converting back to toolkit and ensuring the structure is the
