+++ pyproject.toml
@@ -17,6 +17,7 @@ DJANGO_SETTINGS_MODULE = "simmate.configuration.django.settings_test"
 markers = [
     "blender: mark test as requiring blender.",
     "pymatgen: mark test as pymatgen-compatibility test",
+    "vasp: mark test as requiring vasp.",
 ]
 
 # By default, we only want to run unmarked tests. The simplest way to do this
@@ -25,7 +26,7 @@ markers = [
 # migration folders.
 # I manually remove -m when testing coverage, but am unsure if there's a better
 # way to do this.
-addopts = "--no-migrations -m 'not blender and not pymatgen'"
+addopts = "--no-migrations -m 'not blender and not pymatgen and not vasp'"
 
 # There are a number of warnings that are expected when running our tests.
 # We remove these from our output for clarity.
