+++ src/simmate/workflows/test/test_all_workflow_runs.py
@@ -100,5 +100,5 @@ def test_all_workflow_runs(tmpdir, sample_structures):
         "population-analysis.vasp.badelf-matproj",
     ]
 
-    # for breaking loops
-    [w.nflows_submitted for w in all_flows]
+    # BUG: breaking loops
+    [get_workflow(w).nflows_submitted for w in all_flows]
