+++ src/simmate/workflows/test/test_all_workflow_runs.py
@@ -1,7 +1,5 @@
 # -*- coding: utf-8 -*-
 
-# -*- coding: utf-8 -*-
-
 import pytest
 
 from simmate.workflows.utilities import (
