+++ src/simmate/calculators/vasp/workflows/diffusion/test/test_all_paths.py
@@ -11,6 +11,7 @@ from simmate.calculators.vasp.workflows.diffusion.all import (
 
 
 @pytest.mark.django_db
+@pytest.mark.prefect_bug
 def test_neb(sample_structures, tmpdir, mocker):
     copy_test_files(
         tmpdir,
