+++ src/simmate/workflow_engine/test/test_supervised_staged_shell_task.py
@@ -92,6 +92,7 @@ def test_s3task_methods():
     assert state.is_completed()
     assert os.path.exists(result["directory"])
     os.rmdir(result["directory"])
+    # BUG: see https://github.com/jacksund/simmate/issues/187
 
 
 def test_s3task_1():
