+++ src/simmate/calculators/vasp/workflows/diffusion/test/test_all_paths.py
@@ -11,7 +11,9 @@ from simmate.calculators.vasp.workflows.diffusion.all import (
 
 
 @pytest.mark.django_db
+@pytest.mark.prefect_bug
 def test_neb(sample_structures, tmpdir, mocker):
+
     copy_test_files(
         tmpdir,
         test_directory=__file__,
@@ -42,26 +44,5 @@ def test_neb(sample_structures, tmpdir, mocker):
     )
     assert state.is_completed()
 
-    # BUG: If this job hangs, this line can break that can have it exit
-    # the pytest command successfully. Still trying to figure out why.
+    # BUG: see https://github.com/jacksund/simmate/issues/187
     Diffusion__Vasp__NebAllPaths.nflows_submitted
-    #
-    # As a minimal example, I can get a test to hang in a lone test using...
-    #
-    # from prefect import flow, task
-    # from simmate.database import connect  # causes the hanging test
-    # @task
-    # def dummy_task_1(a):
-    #     return 1
-    # @task
-    # def dummy_task_2(a):
-    #     return 2
-    # @flow
-    # def run_config(source=None, structure=None, **kwargs):
-    #     x = dummy_task_1(source)
-    #     y = dummy_task_2(structure)
-    #     return x.result() + y.result()
-    # def test_workflow():
-    #     state = run_config()
-    #     assert state.is_completed()
-    #     assert state.result() == 3
