+++ src/simmate/workflow_engine/test/test_workflow.py
@@ -66,8 +66,8 @@ def test_workflow():
 
     # test cloud properties
 
-    deployment_id = DummyFlow.deployment_id
-    assert isinstance(deployment_id, str)
+    # deployment_id = DummyFlow.deployment_id
+    # assert isinstance(deployment_id, str)
     # we dont check the actual value bc its randomly generated
 
     n = DummyFlow.nflows_submitted
@@ -76,6 +76,7 @@ def test_workflow():
     # BUG: see https://github.com/jacksund/simmate/issues/187
 
 
+@pytest.mark.prefect_bug  # occassionally fails due to connect timeout
 @pytest.mark.django_db
 def test_workflow_cloud(mocker, sample_structures):
 
