### 2022-07-13 9109e2c
attempt hanging bug fix
