+++ src/simmate/workflow_engine/test/test_workflow.py
@@ -34,7 +34,7 @@ class DummyProject__DummyCaclulator__DummyPreset(Workflow):
 DummyFlow = DummyProject__DummyCaclulator__DummyPreset
 
 
-@pytest.mark.prefect_bug
+@pytest.mark.prefect_db
 def test_workflow():
     # Run the workflow just like you would for the base Prefect class
     flow = DummyFlow.to_prefect_flow()
@@ -65,17 +65,15 @@ def test_workflow():
     DummyFlow.show_parameters()  # a print statment w. nothing else to check
 
     # test cloud properties
-
     deployment_id = DummyFlow.deployment_id
     assert isinstance(deployment_id, str)
     # we dont check the actual value bc its randomly generated
 
     n = DummyFlow.nflows_submitted
     assert isinstance(n, int)
-    # we dont check the actual value bc it could be affected by other parallel tests
-    # BUG: see https://github.com/jacksund/simmate/issues/187
 
 
+@pytest.mark.prefect_db
 @pytest.mark.django_db
 def test_workflow_cloud(mocker, sample_structures):
 
