+++ src/simmate/workflows/test/test_all_workflow_runs.py
@@ -9,7 +9,7 @@ from simmate.workflows.utilities import (
 
 
 @pytest.mark.vasp
-@pytest.mark.prefect_bug
+@pytest.mark.prefect_db
 @pytest.mark.django_db
 def test_all_workflow_runs(tmpdir, sample_structures):
 
@@ -105,6 +105,3 @@ def test_all_workflow_runs(tmpdir, sample_structures):
         "diffusion.vasp.neb-single-path",
         "population-analysis.vasp.badelf-matproj",
     ]
-
-    # BUG: see https://github.com/jacksund/simmate/issues/187
-    [get_workflow(w).nflows_submitted for w in all_flows]
