+++ src/simmate/workflow_engine/workflow.py
@@ -499,7 +499,22 @@ class Workflow:
     async def deployment_id(cls) -> str:
         """
         Grabs the deployment id from the prefect database if it exists, and
-        if not, creates the depolyment
+        if not, creates the depolyment and then returns the id.
+
+        This is a synchronous and cached version of `_get_deployment_id` and
+        this is the preferred method to use for beginners.
+        """
+        return await cls._get_deployment_id()
+
+    @classmethod
+    async def _get_deployment_id(cls) -> str:
+        """
+        Grabs the deployment id from the prefect database if it exists, and
+        if not, creates the depolyment and then returns the id.
+
+        This is an asynchronous method and should only be used when within
+        other async methods. Beginners should instead use the `deployment_id`
+        property.
         """
 
         async with get_client() as client:
@@ -512,19 +527,19 @@ class Workflow:
         # If this is the first time accessing the deployment id, we will need
         # to create the deployment
         if not response:
-            deployment_id = cls._create_deployment()
-            return deployment_id
+            deployment_id = await cls._create_deployment()
 
         # there should only be one deployment associated with this workflow
         # if it's been deployed already.
         elif len(response) == 1:
-            return str(response[0].id)
+            deployment_id = str(response[0].id)
 
         else:
             raise Exception("There are duplicate deployments for this workflow!")
 
+        return deployment_id
+
     @classmethod
-    @async_to_sync
     async def _create_deployment(cls) -> str:
         """
         Registers this workflow to the prefect database as a deployment.
@@ -584,7 +599,7 @@ class Workflow:
 
         async with get_client() as client:
             response = await client.create_flow_run_from_deployment(
-                deployment_id=cls.deployment_id,
+                deployment_id=await cls._get_deployment_id(),
                 **kwargs,
             )
 
