+++ src/simmate/calculators/vasp/workflows/diffusion/test/test_all_paths.py
@@ -10,8 +10,8 @@ from simmate.calculators.vasp.workflows.diffusion.all import (
 )
 
 
+@pytest.mark.prefect_db
 @pytest.mark.django_db
-@pytest.mark.prefect_bug
 def test_neb(sample_structures, tmpdir, mocker):
 
     copy_test_files(
@@ -43,6 +43,3 @@ def test_neb(sample_structures, tmpdir, mocker):
         directory=str(tmpdir),
     )
     assert state.is_completed()
-
-    # BUG: see https://github.com/jacksund/simmate/issues/187
-    Diffusion__Vasp__NebAllPaths.nflows_submitted
