### 2022-07-13 #4364 @bocklund
@rkurchin @utf  Yes, sorry for the delay. My review is complete and I can recommend accepting Simmate into JOSS. Very nice work @jacksund!
