+++ src/simmate/workflow_engine/workflow.py
@@ -1,5 +1,269 @@
 # -*- coding: utf-8 -*-
 
+"""
+This module defines the base class for all workflows in Simmate. When learning 
+how use workflows, make sure you have gone through our intro 
+[tutorials](https://github.com/jacksund/simmate/tree/main/tutorials). You
+can then read through these guides for more features.
+
+It useful to know that our Worklfow class builds Prefect Flows under the 
+hood, so having knowledge of Prefect can be useful in advanced cases. For 
+advanced use or when building new features, we recommend going through the
+prefect tutorials located
+[here](https://orion-docs.prefect.io/tutorials/first-steps/).
+
+
+
+# Using existing workflows
+
+
+## Loading a workflow
+
+To import a specific workflow, see the `simmate.workflows` module. For all the
+example uses below, we will look at the `static-energy.vasp.matproj` workflow,
+which we load with:
+    
+``` python
+from simmate.workflows.all import StaticEnergy__Vasp__Matproj as workflow
+```
+
+
+## Viewing input parameters
+
+
+
+## Running a workflow (local)
+
+To run a workflow locally (i.e. directly on your current computer), you can
+use the `run` method. This is returns a Prefect `State` object, which allows
+you to check if the run completed successfully or not. Then final output of your
+workflow run can be accessed using `state.result()`. You can read more about
+`State` objects [here](https://orion-docs.prefect.io/concepts/states/), but as
+a quick example:
+
+``` python
+state = workflow.run(
+    structure="NaCl.cif", 
+    command="mpirun -n 4 vasp_std > vasp.out",
+)
+
+if state.is_completed():  # optional check
+    result = state.result()
+```
+
+Outside of python, you can also run workflows from the command line, using YAML 
+files, or the website interface. These approaches are covered in the intro
+tutorials.
+
+
+## Running a workflow (cloud)
+
+When you want to schedule a workflow to run elsewhere, you must first make sure
+you have your computational resources configured. You can then run workflows
+using the `run_cloud` method, which returns a Prefect flow run id.
+
+``` python
+prefect_flow_run_id = workflow.run_cloud(
+    structure="NaCl.cif", 
+    command="mpirun -n 4 vasp_std > vasp.out",
+)
+```
+
+## Accessing results
+
+In addition to using `state.result()` as described above, you can also view
+results database through the `database_table` attribute (if one is available).
+This returns a Simmate database object for results of ALL runs of this workflow.
+Guides for filtering and manulipating the data in this table is covered in
+the `simmate.database` module guides. But as an example:
+
+``` python
+table = workflow.database_table
+
+# pandas dataframe that you can view in Spyder
+df = table.obects.to_dataframe()
+
+# or grab a specific run result and convert to a toolkit object
+entry = table.objects.get(prefect_flow_run_id="example-123456")
+structure = entry.to_toolkit()
+```
+
+<!--
+TODO:
+    - viewing parameter names
+    - accessing results (database_table)
+    - run_cloud and other cloud methods
+    - different documentation levels (parameters, mini docstring, etc.)
+    - finding the same workflow in the website UI (place this )
+-->
+
+# Creating new workflows
+
+## Naming workflows (required)
+
+Higher level features such as the website interface require that workflow
+names follow a certain format. If you skip this step, your workflows
+will fail and cause errors elsewhere.
+
+First, we need to update the workflow name to match Simmate's naming
+conventions, which includes:
+    1.  The type of analysis the workflow is doing
+    2.  The "calculator" (or program) that the workflow uses to run
+    3.  A unique name to identify the settings used
+
+Examples for each part would be:
[…]
