+++ tutorials/02_ Run_a_workflow.md
@@ -37,7 +37,7 @@ direct
 ```
 3. View a list of all workflows available with `simmate workflows list-all`
 4. Interactively learn about all workflows with `simmate workflows explore`
-5. View the settings used for the `static-energy/mit` workflow with `simmate workflows show-config static-energy/mit`
+5. View the settings used for the `static-energy.vasp.mit` workflow with `simmate workflows show-config static-energy.vasp.mit`
 6. Copy and paste VASP POTCAR files to the folder `~/simmate/vasp/Potentials`. Be sure to unpack the `tar.gz` files. This folder will have the potentials that came with VASP -- and with their original folder+file names:
 ```
 # Located at /home/my_username (~)
@@ -62,18 +62,18 @@ simmate/
 
 - **OPTION 1**: Define all settings directly in the command line (best for quick submitting and testing)
 ``` bash
-simmate workflows run static-energy/mit --structure POSCAR
+simmate workflows run static-energy.vasp.mit --structure POSCAR
 
 # OR
 
-simmate workflows run static-energy/mit --structure POSCAR --command "mpirun -n 5 vasp_std > vasp.out"
+simmate workflows run static-energy.vasp.mit --structure POSCAR --command "mpirun -n 5 vasp_std > vasp.out"
 ```
 
 - **OPTION 2**: Run from a settings file in yaml format (best for complex settings)
 ``` yaml
 # In a file named "my_example.yaml".
 # Note, different workflows accept different settings here.
-workflow_name: static-energy/mit
+workflow_name: static-energy.vasp.mit
 structure: POSCAR
 command: mpirun -n 5 vasp_std > vasp.out  # OPTIONAL
 directory: my_new_folder  # OPTIONAL
@@ -206,7 +206,7 @@ loop_
   Cl  Cl1  1  0.50000000  0.50000000  0.50000000  1
 ```
 
-Nearly all files that you will interact with are text files -- just in different formats. That's where file extensions come in. They indicate what format we are using. Files named `something.cif` just tell programs we have a text file written if the CIF structure format. VASP uses the name POSCAR (without any file extension) to show its format. So rename your file from `POSCAR.txt` to `POSCAR`, and now all programs (VESTA, OVITO, and others) will know what to do with your structure. In Windows, you will often receive a warning about changing the file extension. Ignore the warning and change the extension.
+Nearly all files that you will interact with are text files -- just in different formats. That's where file extensions come in. They indicate what format we are using. Files named `something.cif` just tell programs we have a text file written in the CIF structure format. VASP uses the name POSCAR (without any file extension) to show its format. So rename your file from `POSCAR.txt` to `POSCAR`, and now all programs (VESTA, OVITO, and others) will know what to do with your structure. In Windows, you will often receive a warning about changing the file extension. Ignore the warning and change the extension.
 
 If you're using the command-line to create/edit this file, you can use the copy (`cp`) command to make the `POSCAR` file:
 
@@ -222,7 +222,7 @@ We now have our structure ready to go! Let's get back to running a workflow.
 
 ## Configuring Potentials (for VASP users)
 
-> :warning: once Simmate switches from VASP to a free DFT alternative, this section of the tutorial will be removed. 
+> :warning: once Simmate switches from VASP to a free DFT alternative, this section of the tutorial will be removed.
 
[…]
