+++ src/simmate/website/workflow_execution/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class WorkflowExecutionConfig(AppConfig):
-
-    # use the full import path for this app b/c it's within a package
-    name = "simmate.website.workflow_execution"
