+++ src/simmate/workflow_engine/execution/worker.py
@@ -1,169 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import time
-
-# import pickle
-import cloudpickle  # needed to serialize Prefect workflow runs and tasks
-
-from django.db import transaction
-
-from simmate.workflow_engine.execution.database import WorkItem
-
-# In the future, I might want this worker to be a local dask cluster that has
-# two threads going. One thread will update the queue database with a "heartbeat" to
-# let it know that it is still working on tasks. The other thread will run the
-# given workitems in serial. I could even allow for more threads so that this
-# worker can run multiple workitems at once and in parallel. I'm not sure if
-# asyncio would play into this though.
-
-# This string is just something fancy to display in the console when a worker
-# starts up.
-# This uses "Small Slant" from https://patorjk.com/software/taag/
-HEADER_ART = r"""
-   _____                  __        _      __         __
-  / __(_)_ _  __ _  ___ _/ /____   | | /| / /__  ____/ /_____ ____
- _\ \/ /  ' \/  ' \/ _ `/ __/ -_)  | |/ |/ / _ \/ __/  '_/ -_) __/
-/___/_/_/_/_/_/_/_/\_,_/\__/\__/   |__/|__/\___/_/ /_/\_\\__/_/
-
-"""
-
-
-class SimmateWorker:
-    def __init__(
-        self,
-        # limit of tasks and lifetime of the worker
-        nitems_max=None,
-        timeout=None,
-        # wait_on_timeout=False, # TODO
-        # settings on what to do when the queue is empty
-        close_on_empty_queue=False,
-        waittime_on_empty_queue=60,
-    ):
-
-        # the maximum number of workitems to run before closing down
-        # if no limit was set, we can go to infinity!
-        self.nitems_max = nitems_max or float("inf")
-
-        # Don't start a new workitem after this time. The worker will be shut down.
-        # if no timeout was set, use infinity so we wait forever.
-        self.timeout = timeout or float("inf")
-
-        # whether to wait on the running workitems to finish before shutting down
-        # the timedout worker.
-        # self.wait_on_timeout = wait_on_timeout # # TODO
-
-        # whether to close if the queue is empty
-        self.close_on_empty_queue = close_on_empty_queue
-        # if the queue is found to be empty, we will give it one last chance
-        # to fill. Check the queue again after this time sleeping and if it is
-        # still empty, close the worker.
-        self.waittime_on_empty_queue = waittime_on_empty_queue
-
-    def start(self):
-
-        # print the header in the console to let the user know the worker started
-        print(HEADER_ART)
-
-        # establish starting point for the worker
-        time_start = time.time()
-        ntasks_finished = 0
-
-        # Loop endlessly until one of the following happens...
-        #   the timeout limit is hit
-        #   the queue is empty
-        #   the nitems limit is hit
-        while True:
-
-            # check for timeout before starting a new workitem and exit
-            # if we've hit the limit.
-            if (time.time() - time_start) > self.timeout:
-                # TODO - check wait_on_timeout if running in parallel.
-                print("The time-limit for this worker has been hit.")
-                return
-
-            # check the number of jobs completed so far, and exit if we hit
-            # the limit
-            if ntasks_finished >= self.nitems_max:
-                print("Maxium number of WorkItems hit for this worker.")
-                return
-
-            # check the length of the queue and while it is empty, we want to
-            # loop. The exception of looping endlessly is if we want the worker
-            # to shutdown instead.
-            while self.queue_size() == 0:
-                # if it is empty, we want to sleep for a little and check again
-                time.sleep(self.waittime_on_empty_queue)
-
-                # This is a special condition where we may want to close the
-                # worker if the queue stays empty
[…]
