+++ src/simmate/configuration/example_project/example_app/tests.py
@@ -1,9 +1,15 @@
 # -*- coding: utf-8 -*-
 
 """
-This file is important only when your project grows and starts getting users.
+This file is important as your project grows and starts getting users.
 Here, you need to make sure any changes you make don't break old features and
 code.
+
+Simmate does not have full guides on writing unittests yet, so we recommend
+looking through other packges such as...
+- [pytest](https://docs.pytest.org/en/7.1.x/)
+- [pytest-django](https://pytest-django.readthedocs.io/en/latest/)
+- [django test tutorials](https://docs.djangoproject.com/en/4.0/intro/tutorial05/)
 """
 
 from django.test import TestCase
