+++ tutorials/06_Building_custom_workflows.md
@@ -79,24 +79,23 @@ Both of these approaches are only suitable for customizing settings for a few ca
 
 ## Create new & advanced workflows
 
-1. Simmate defines a base `Workflow` class to help with common material science analyses, but at its core, it works exactly the same as Prefect workflows. Therefore, you can take advantage of all of [Prefect's guides](https://docs.prefect.io/core/) when building new workflows. The simplest possible workflow can look something like...
+1. Simmate defines a base `Workflow` class to help with common material science analyses, but at its core, it uses Prefect Flows. Therefore, if you don't care for advanced Simmate features, you can build custom workflows directly with Prefect. Take advantage of all of [Prefect's guides](https://orion-docs.prefect.io/tutorials/first-steps/) when building new workflows. The simplest possible workflow can look something like...
 
 ``` python
-from simmate.workflow_engine import Workflow, task
+from prefect import flow
 
-@task
-def hello_task():
-    print("Hello World!")
+@flow
+def my_favorite_function():
+    print("This function doesn't do much")
+    return 42
 
-with Workflow("hello-flow") as my_new_workflow:
-    hello_task()
-
-my_new_workflow.run()
+state = my_favorite_function()
+result = state.result()
 ```
 
-2.  In building a material science workflow, we may want two common steps: (a) running a calculation through a program like VASP and (b) saving the results to our database. We must therefore define individual `Task`s as well as a `DatabaseTable` to store results in. The next steps will address these.
+2. However, we've built many helpful features on top of Simmate to make your workflows more powerful and easier to write. In building a material science workflow, we may want two common steps: (i) running a calculation through a program like VASP and (ii) saving the results to our database. We must therefore define individual `Task`s as well as a `DatabaseTable` to store results in. The next steps will address both of these.
 
-3. To build a simple task, we use the `@task` decorator like we did above with our `hello_task` function. However, because material science frequently involves calling external programs, we've created a powerful `S3Task` class to inherit from. A `S3Task` class and implement their own `setup`, `execute`, and `workup` methods, and even add functions for error handling. Here's how you would build a custom `S3Task` from scratch:
+3. Because material science frequently involves calling external programs, we've created a powerful `S3Task` class to inherit from. A `S3Task` class can implement it's own `setup`, `execute`, and `workup` methods, and even add functions for error handling. Here's how you would build a custom `S3Task` from scratch:
 
 ``` python
 import os
@@ -107,13 +106,15 @@ from simmate.workflow_engine import S3Task
 
 class ExampleTask(S3Task):
 
-    command = "echo 'I am calling some program here!' > output.txt" 
-
-    def setup(self, structure, directory, **kwargs):
+    command = "echo 'I am calling some program here!' > output.txt"
+    
+    @staticmethod
+    def setup(directory, structure, **kwargs):
         filename = os.path.join(directory, "my_input_structure.cif")
         structure.to("cif", filename)
 
-    def workup(self, directory: str):
+    @staticmethod
+    def workup(directory):
         # do some workup or analysis on the output files. Here we just return
         # the original structure back as the "results"
         filename = os.path.join(directory, "my_input_structure.cif")
@@ -132,19 +133,14 @@ nacl_example = Structure(
     coords=[[0.0, 0.0, 0.0], [0.5, 0.5, 0.5]],
 )
 
-task = ExampleTask()
-
-results = task.run(
-    structure=nacl_example,
-    # command="echo "my new command" > output.txt,  # OPTIONAL
-    # directory="Test-Folder",  # OPTIONAL
-)
+# Tasks classes don't need to be initialized. You can call them directly.
+state = ExampleTask.run(structure=nacl_example)
+result = state.result()
 
 # The output is a dictionary containing the return value from the workup 
 # method along with other information about the task run.
[…]
