### 2022-07-17 6113658
:books: update docs and unittests (#189)
* update workflow and s3task docs

* begin updating tutorials

* remove django executor

* update tutorials and example project

* run black
