+++ src/simmate/website/workflow_execution/urls.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from django.urls import path
-from . import views
-
-app_name = "workflow_execution"
-urlpatterns = []
