+++ src/simmate/workflow_engine/supervised_staged_shell_task.py
@@ -1,5 +1,116 @@
 # -*- coding: utf-8 -*-
 
+"""
+# The Supervised-Staged-Shell Task (aka "S3Task")
+
+This class contains the core functionality to **supervise** a **staged** task
+involving some **shell** command.
+
+Let's breakdown what this means...
+
+A *shell* command is a single call to some external program. For example,
+VASP requires that we call the "vasp_std > vasp.out" command in order to run a
+calculation. We consider calling external programs a *staged* task made
+up of three steps:
+
+- setup = writing any input files required for the program
+- execution = actually calling the command and running our program
+- workup = loading data from output files back into python
+
+And for *supervising* the task, this means we monitor the program while the
+execution stage is running. So once a program is started, Simmate can check
+output files for common errors/issues. If an error is found, we stop the
+program, fix the issue, and then restart it. Any fixes that were made are
+written to "simmate_corrections.csv".
+
+
+<!--
+TODO: Make a simple diagram to visualize the overall process and add it here.
+It will be similar to Custodian's, but we don't have a list of jobs here.
+https://materialsproject.github.io/custodian/index.html#usage
+The steps are...
+
+- Write Input Files based on custom+defualt settings
+- Run the calculation by calling the program
+- Load ouput files
+- check for errors
+- [correct them, rerun]
+- postprocess/analysis
+-->
+
+This entire process (the different stages and monitoring) is carried out
+using the `run` method. You rarely use this class directly. Instead,
+you typically use a subclass of it. As a user, you really just need to do
+something like this:
+
+``` python
+from simmate.calculator.example.tasks import ExampleTask
+
+my_task = ExampleTask()
+my_result = my_task.run()
+```
+
+And that's it!
+
+For experts, this class can be viewed as a combination of prefect's ShellTask,
+a custodian Job, and Custodian monitoring. When subclassing this, we can absorb
+functionality of `pymatgen.io.vasp.sets` too. By merging all of these together
+into one class, we make things much easier for users and creating new Tasks.
+
+
+# Building a custom S3task
+
+This class is commonly used to make tasks for our calculator modules, so you
+will likely want to subclass this. Here is a basic example of inheriting
+and then running a task:
+
+``` python
+
+from simmate.workflow_engine import S3Task
+from example.error_handlers import PossibleError1, PossibleError2
+
+
+class ExampleTask(SSSTask):
+
+    command = "echo example"  # just prints out "example"
+    
+    # settings for error handling
+    max_corrections = 7
+    error_handlers = [PossibleError1, PossibleError2]
+    polling_timestep = 0.1
+    monitor_freq = 10
+    
+    # custom attributes
+    some_new_setting = 123
+
+    @classmethod
+    def setup(cls, directory, custom_parmeter, **kwargs):
+        print("I'm setting things up!")
+        print(f"My new setting value is {cls.some_new_setting}")
+        print(f"My new parameter value is {custom_parmeter}")
+
+    @staticmethod
+    def workup(directory):
+        print("I'm working things up!")
+
+
+task = ExampleTask()
+result = task.run()
+```
+
+There are a two important things to note here:
+
+1. It's optional to write new `setup` or `workup` methods. But if you do...
+    - Both `setup` and `workup` method should be either a staticmethod or classmethod
+    - Custom `setup` methods require the `directory` and `**kwargs` input parameters.
+    - Custom `workup` methods require the `directory` input paramter
+2. It's optional to set/overwrite attributes. You can also add new ones too.
+
+For a full (and advanced) example, of a subclass take a look at
+`simmate.calculators.vasp.tasks.base.VaspTask` and the tasks that use it like
[…]
