+++ src/simmate/workflow_engine/__init__.py
@@ -9,6 +9,7 @@ __doc__ = get_doc_from_readme(__file__)
 # module, so it may be worth isolating them for faster imports.
 from .error_handler import ErrorHandler
 from .supervised_staged_shell_task import S3Task
+from .worker import Worker
 
 # All imports below this point depend on the simmate.database module and therefore
 # need a database connection. This line sets up the database connection so
