### 2022-07-17 03dfaa0
update start-worker command
