### 2022-07-18 jacksund/simmate#191 @utf
Add warning about disk space needed to install simmate
A fresh installation of simmate on OSX (following the instructions in the installation document) comes to over 1.2 GB of data. This might come as a shock to people used to running small python scripts for orchestrating calculations and could take up a reasonable amount of home space of computing clusters. It might be nice to add a warning about this in the documentation.

https://github.com/openjournals/joss-reviews/issues/4364
