### 2022-07-18 jacksund/simmate#192 @utf
`simmate workflows show-config static-energy.vasp.mit` command cannot find workflow
In the "Run a workflow" docs, the user is suggested to run the command `simmate workflows show-config static-energy.vasp.mit` however this results in the output.

```
LOADING WORKFLOW...
Traceback (most recent call last):
  File "/Users/alexmganose/.conda/envs/simmate/bin/simmate", line 10, in <module>
    sys.exit(simmate())
  File "/Users/alexmganose/.conda/envs/simmate/lib/python3.10/site-packages/click/core.py", line 1130, in __call__
    return self.main(*args, **kwargs)
  File "/Users/alexmganose/.conda/envs/simmate/lib/python3.10/site-packages/click/core.py", line 1055, in main
    rv = self.invoke(ctx)
  File "/Users/alexmganose/.conda/envs/simmate/lib/python3.10/site-packages/click/core.py", line 1657, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
  File "/Users/alexmganose/.conda/envs/simmate/lib/python3.10/site-packages/click/core.py", line 1657, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
  File "/Users/alexmganose/.conda/envs/simmate/lib/python3.10/site-packages/click/core.py", line 1404, in invoke
    return ctx.invoke(self.callback, **ctx.params)
  File "/Users/alexmganose/.conda/envs/simmate/lib/python3.10/site-packages/click/core.py", line 760, in invoke
    return __callback(*args, **kwargs)
  File "/Users/alexmganose/.conda/envs/simmate/lib/python3.10/site-packages/simmate/command_line/workflows.py", line 180, in show_config
    workflow = get_workflow(
  File "/Users/alexmganose/.conda/envs/simmate/lib/python3.10/site-packages/simmate/workflows/utilities.py", line 142, in get_workflow
    raise ModuleNotFoundError(
ModuleNotFoundError: The workflow you provided isn't known. Make sure you don't have any typos! If you want a list of all available workflows, use the command `simmate workflows list-all`. You can also interactively explore workflows with `simmate workflows explore`
```

Specifically, the command is written here:https://github.com/jacksund/simmate/blob/6113658945233da6082e10304aa9693a2cfd8272/tutorials/02_%20Run_a_workflow.md?plain=1#L40

https://github.com/openjournals/joss-reviews/issues/4364
