### 2022-07-18 jacksund/simmate#192 @utf
Investigating in more detail, the subsequent commands that use the same workflow name will also fail.

I believe the correct workflow name should be `static-energy/mit `
