### 2022-07-18 jacksund/simmate#192 @jacksund
This is because the workflow names are changed in the upcoming release. Change your branch to the v0.6.0 if that's what you have installed.

This does get to a bigger issue is that I need to separate dev version tutorials from the most recent release.
