### 2022-07-18 jacksund/simmate#192 @jacksund
I put a warning at the top of the first tutorial for this issue [here](https://github.com/jacksund/simmate/blob/main/tutorials/01_Installation.md), but that might not be enough
