### 2022-07-18 jacksund/simmate#191 @jacksund
The 1.2GB you metnioned includes the anaconda installation + all simmate dependencies, right? I believe simmate itself is under 2 MB when unpacked (based on the [pypi download](https://pypi.org/project/simmate/#files)). Just want to make sure there isn't something else being downloaded -- such as the test files.

I can add this warning though :smile:
