### 2022-07-18 #4364 @utf
Apologies for the delay getting back my review.

I've finished reviewing simmate and recommend that it be accepted. There is an incredible amount of functionality in the package, the documentation is very thorough and easy to follow, and I can confirm that all examples work as intended.
