### 2022-07-18 #4364 @rkurchin
Some very minor editorial comments (this is by far the smallest number of these I've ever had, so 👏 from my editor hat 🎩):
* line 41: when used as a noun, "best practices" doesn't need to be hyphenated
* 49: "full-feature" -> "full-featured"
* 42: "high-level" is a compound adjective and _should_ be hyphenated here
* 66: needs a comma after "integrations"
