+++ paper.md
@@ -33,11 +33,11 @@ Over the past decade, the automation of electronic structure codes has led to ma
 
 Simmate strives to simplify the process for researchers who are setting up a full-featured server. For the purposes of beginners, we desired a framework that could run locally without requiring any additional setup. For the purposes of experts, we sought to enable scaling of calculations across any number of resources and to facilitate the addition of new functionality. Simmate accomplishes these goals by (i) building on top of popular, well-established packages for workflow orchestration, database management, and materials science analysis and (ii) distributing our software as an "all-in-one" package that integrates all of these features. This is contrary to analogous software ecosystems that maintain custom packages and distribute solutions as separate programs.
 
-For example, while other materials science ecosystems write workflow managers and task distribution from scratch (e.g. Fireworks [@fireworks] or AiiDA [@aiida]), we instead use high-level, beginner-friendly packages such as Dask [@dask] and Prefect [@prefect]. Using well-established packages also extends to our choice of website framework (Django [@django]) and underlying materials analysis toolkit (PyMatGen [@pymatgen]). The use of popular packages lets Simmate users take advantage of these packages' large user communities, abundant guides, and robust coding standards, while Simmate handles the integration of these packages in the context of materials science. This greatly facilitates the addition of new features while also enabling best-practices.
+For example, while other materials science ecosystems write workflow managers and task distribution from scratch (e.g. Fireworks [@fireworks] or AiiDA [@aiida]), we instead use high-level, beginner-friendly packages such as Dask [@dask] and Prefect [@prefect]. Using well-established packages also extends to our choice of website framework (Django [@django]) and underlying materials analysis toolkit (PyMatGen [@pymatgen]). The use of popular packages lets Simmate users take advantage of these packages' large user communities, abundant guides, and robust coding standards, while Simmate handles the integration of these packages in the context of materials science. This greatly facilitates the addition of new features while also enabling best practices.
 
-To understand our aim to be an “all-in-one” solution, it is useful to compare Simmate with the collection of packages developed by the Materials Project [@matproj]. The Materials Project is powered by many smaller packages, each with a specific use case – e.g. Atomate for a workflow library [@atomate], Fireworks for workflow orchestration [@fireworks], Custodian for error handling [@pymatgen], EMMET for schemas/APIs [@emmet], MPContribs for third-party data [@mpcontrib], and many others. Each of these are powerful tools, but it can require significant effort and expertise to properly integrate several packages into a full-feature server. Meanwhile, Simmate contains modules for each of these features within a single, larger package – e.g. within our `workflows`, `workflow_engine`, `database`, and `website` modules. By maintaining these features within a single space, high level integrated features can more readily be developed. This includes features that are unique to Simmate, such as dynamically built REST APIs and website interfaces. There is also a `run-server` command that compiles all features (including user-defined projects) with minimal setup. Many comparisons can be made between Simmate modules and existing packages from the materials science community, but most notably, Simmate focuses on the unification of components for high-level features and capabilities.
[…]
