+++ pyproject.toml
@@ -28,7 +28,7 @@ markers = [
 # migration folders.
 # I manually remove -m when testing coverage, but am unsure if there's a better
 # way to do this.
-addopts = "--no-migrations -m 'not blender and not pymatgen and not vasp and not slow'"
+addopts = "--no-migrations -m 'not blender and not pymatgen and not vasp and not slow and not prefect_db'"
 
 # There are a number of warnings that are expected when running our tests.
 # We remove these from our output for clarity.
