### 2022-07-19 #4364 @jacksund
@rkurchin awesome, thanks for editing the manuscript! 🥳 I've addressed your comments in the newest version.

Also, I'll make a new release + create a DOI on Zennedo once I get the chance.  There was a libwebp release made a few hours ago that prevents any conda-forge updates and causes my CI for MacOS to fail. They already have [an issue](https://github.com/conda-forge/libwebp-feedstock/issues/40) and PR open, so it should hopefully have a fix soon.
