### 2022-07-19 92ac9b0
:seedling: add start-worker to cli (#190)
* add simmate worker

* update start-worker command

* add warning for disk space requirement (#193)

* add simmate worker

* update start-worker command

* Update CHANGELOG.md

* Update pyproject.toml
