+++ src/simmate/workflow_engine/worker.py
@@ -1,0 +1,265 @@
+# -*- coding: utf-8 -*-
+
+import anyio
+import time
+import uuid
+from functools import cached_property
+
+from prefect.client import get_client
+from prefect.exceptions import PrefectHTTPStatusError
+from prefect.agent import OrionAgent
+from prefect.settings import PREFECT_AGENT_QUERY_INTERVAL
+from prefect.orion.schemas.filters import FlowRunFilter  # FlowFilter,
+
+from simmate.utilities import async_to_sync
+
+# This string is just something fancy to display in the console when a worker
+# starts up.
+# This uses "Small Slant" from https://patorjk.com/software/taag/
+HEADER_ART = r"""
+   _____                  __        _      __         __
+  / __(_)_ _  __ _  ___ _/ /____   | | /| / /__  ____/ /_____ ____
+ _\ \/ /  ' \/  ' \/ _ `/ __/ -_)  | |/ |/ / _ \/ __/  '_/ -_) __/
+/___/_/_/_/_/_/_/_/\_,_/\__/\__/   |__/|__/\___/_/ /_/\_\\__/_/
+
+"""
+
+
+class Worker:
+    """
+    A worker is a process that checks the Prefect database for scheduled workflow
+    runs and then submits them.
+
+    Workers are only required when you are using a workflow's `run_cloud` method.
+
+    This is effectively a combination of a Prefect Agent and Queue, where we
+    configure default values for a queue and automatically perform the setup.
+    """
+
+    def __init__(
+        self,
+        queue_name: str = None,
+        concurrency_limit: int = 1,
+        nflows_max: int = None,
+        timeout: float = None,
+        close_on_empty_queue: bool = False,
+    ):
+        # This will be used when creating the queue
+        self.concurrency_limit = concurrency_limit
+
+        # if no name was given, we generate a random one
+        if not queue_name:
+            # generate a unique name for the worker, where we use uuid and take
+            # the first 8 values (the rest is overkill for this application)
+            unique_id = str(uuid.uuid4())[:8]
+            queue_name = f"simmate-worker-{unique_id}"
+
+        self.queue_name = queue_name
+
+        # Grab the worker queue.
+        # This is a cached property, so simply calling it and asserting that
+        # it exists is sufficient to set everything up.
+        assert self.queue_id
+
+        # the maximum number of workitems to run before closing down
+        # if no limit was set, we can go to infinity!
+        self.nflows_max = nflows_max or float("inf")
+
+        # Don't start a new workitem after this time. The worker will be shut down.
+        # if no timeout was set, use infinity so we wait forever.
+        self.timeout = timeout or float("inf")
+
+        # whether to close if the queue is empty
+        self.close_on_empty_queue = close_on_empty_queue
+
+    @async_to_sync
+    async def run(self):
+        """
+        Starts the worker and begins running scheduled workflows.
+        """
+
+        # !!! This method is really a fork of Prefect's `prefect agent start`
+        # command, so a lot of this could be move to a prefect contribution.
+        # https://github.com/PrefectHQ/prefect/blob/orion/src/prefect/cli/agent.py
+
+        # print the header in the console to let the user know the worker started
+        print(HEADER_ART)
+
+        # establish starting point for the worker
+        time_start = time.time()
+        nflows_submitted = 0
+
+        async with OrionAgent(work_queue_name=self.queue_name) as agent:
+
+            print(f"Worker started! Looking for work from queue '{self.queue_name}'...")
+
+            # Loop endlessly until one of the following happens...
+            #   the timeout limit is hit
+            #   the queue is empty
+            #   the nitems limit is hit
+            #   the user stops the process with ctrl+C  # BUG
+            while True:
+
+                # check for timeout before starting a new workitem and exit
[…]
