### 2022-07-19 ccedff9
:grapes: prep release (#195)
* update docs and run pdoc

* Update CHANGELOG.md
