### 2022-07-19 d532e18
fix missing svg files (#196)
