+++ setup.py
@@ -29,7 +29,7 @@ setup(
     name="simmate",
     # Versions should comply with PEP 440:
     # https://www.python.org/dev/peps/pep-0440/
-    version="0.7.0",
+    version="0.7.1",
     # a quick summary and then README
     description="Simmate is a toolbox for computational materials research.",
     long_description=long_description,
@@ -124,6 +124,7 @@ setup(
             "**/*.csv",
             "**/*.yaml",
             "**/*.html",
+            "**/*.svg",
         ],
     },
 )
