### 2022-07-19 #4364 @jacksund
@rkurchin I just made the Zenodo upload [here](https://zenodo.org/record/6863068#.Ytb6pNLMJkg)

v0.7.1 --> DOI: 10.5281/zenodo.6863068
all versions --> DOI: 10.5281/zenodo.6863067
