### 2022-07-19 cc8160f
move from setup to pyproject
