+++ setup.py
@@ -1,130 +1,8 @@
-# -*- coding: utf-8 -*-
 
-"""
-This file establishes the info needed to install simmate via pip and also how to
-upload it to PyPI. This file was copied and editted from the following source:
-    https://github.com/pypa/sampleproject
+# This file is only required for editable installs using pip. We can
+# remove this file once the following issue is closed:
+#   https://github.com/pypa/setuptools/issues/2816
 
-For the details on conda-forge installation see:
-    https://github.com/conda-forge/simmate-feedstock
-"""
+from setuptools import setup
 
-from setuptools import setup, find_packages
-import pathlib
-
-# Grab the full path to this file
-here = pathlib.Path(__file__).parent.resolve()
-
-# Get the long description from the README file
-# long_description = (here / "README.md").read_text(encoding="utf-8")
-# NOTE: I remove the long description and just have users move directly to github
-long_description = (
-    "Please do NOT install Simmate with pip. We recommend conda instead."
-    " Visit our [github page](https://github.com/jacksund/simmate)"
-    " for more information."
-)
-
-setup(
-    # published name for pip install to use
-    name="simmate",
-    # Versions should comply with PEP 440:
-    # https://www.python.org/dev/peps/pep-0440/
-    version="0.7.1",
-    # a quick summary and then README
-    description="Simmate is a toolbox for computational materials research.",
-    long_description=long_description,
-    long_description_content_type="text/markdown",
-    # Link to our homepage. Use github for now.
-    url="https://github.com/jacksund/simmate",
-    # Lead dev info
-    author="Jack D. Sundberg",
-    author_email="jacksund@live.unc.edu",
-    # Classifiers help users find your project by categorizing it.
-    # For a list of valid classifiers, see https://pypi.org/classifiers/
-    classifiers=[
-        "Development Status :: 3 - Alpha",
-        "Intended Audience :: Science/Research",
-        "License :: OSI Approved :: BSD License",
-    ],
-    # Keywords -- Removed for now. Below are PyMatgen's keywors for future reference
-    # keywords=(
-    #     "VASP, gaussian, ABINIT, nwchem, qchem, materials, science, project, "
-    #     "electronic, structure, analysis, phase, diagrams, crystal"
-    # ),
-    # Indicate which directory the source coude is in
-    package_dir={"": "src"},
-    packages=find_packages(where="src"),
-    # Specify which Python versions supported.
-    python_requires=">=3.10",
-    # This field lists other packages that your project depends on to run.
-    install_requires=[
-        # Core dependencies
-        "numpy>=1.22.0",
-        "pandas>=1.3.5",
-        "django>=4.0.0",
-        "prefect==2.0b8",  # !!! FIXED VERSIONING UNTIL FIRST STABLE RELEASE
-        "dask>=2021.12.0",
-        "click>=8.0.3",
-        # Extra (smaller) dependencies & utilities
-        "django-allauth>=0.50.0",  # for website accounts and google/github sign-ins
-        "django-crispy-forms>=1.13.0",  # for formatting of online forms
-        "django-pandas>=0.6.6",  # for converting QuerySets to PandasDataFrames
-        "dask-jobqueue>=0.7.3",  # for submitting on clusters
-        "dj-database-url>=0.5.0",  # for DigitalOcean URL conversion
-        "djangorestframework>=3.13.1",  # for our REST API
-        "django-filter>=21.1",  # sets up automatic filters for our REST API
-        "pyyaml>=6.0",  # for yaml configuration files
-        "plotly>=5.4.0",  # for interactive plots and visualization
-        "tqdm>=4.62.3",  # for monitoring progress of long for-loops
[…]
