+++ pyproject.toml
@@ -1,7 +1,150 @@
 
-# This file configures settings for utilities like pytest and coverage. We prefer 
-# to use this file over files like "pytest.ini" because pyproject.toml has 
-# more uses that we may need in the future (i.e. it can use other modules)
+# This file establishes the info needed to install simmate via pip, how to
+# upload it to PyPI, and how to run python packages like pytest and coverage.
+
+# For more information on this file and how to make similar ones, see:
+#    https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
+#    https://github.com/pypa/sampleproject
+
+# For the details on conda-forge installation see:
+#    https://github.com/conda-forge/simmate-feedstock
+
+# Note, for utilities like pytest and coverage, we prefer to use this file 
+# over files like "pytest.ini" because pyproject.toml has more uses that 
+# we may need in the future (i.e. it can use other modules).
+
+# -----------------------------------------------------------------------------
+
+[build-system]
+# These are the assumed default build requirements from pip:
+# https://pip.pypa.io/en/stable/reference/pip/#pep-517-and-518-support
+requires = ["setuptools>=43.0.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+# -----------------------------------------------------------------------------
+
+[project]
+
+# published name for pip install to use
+name="simmate"
+
+# Versions should comply with PEP 440:
+# https://www.python.org/dev/peps/pep-0440/
+# https://semver.org/
+version="0.7.1"
+# In the future, I may be able to dynamically determine this:
+#   https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html#dynamic-metadata
+
+# Maintainer info
+authors = [{name = "Jack D. Sundberg", email = "jacksund@live.unc.edu"}]
+
+# a quick summary and then README
+# NOTE: instead of our main README, I tell users to move directly to github
+summary="Simmate is a toolbox for computational materials research."
+description_content_type="text/markdown"
+description = "Please do NOT install Simmate with pip. We recommend conda instead. Visit our [github page](https://github.com/jacksund/simmate) for more information."
+
+# Specify which Python versions are supported.
+python_requires=">=3.10"
+
+# This field lists other packages that your project depends on to run.
+dependencies=[
+    # Core dependencies
+    "numpy>=1.22.0",
+    "pandas>=1.3.5",
+    "django>=4.0.0",
+    "prefect==2.0b8",  # !!! FIXED VERSIONING UNTIL FIRST STABLE RELEASE
+    "dask>=2021.12.0",
+    "click>=8.0.3",
+    # Extra (smaller) dependencies & utilities
+    "django-allauth>=0.50.0",  # for website accounts and google/github sign-ins
+    "django-crispy-forms>=1.13.0",  # for formatting of online forms
+    "django-pandas>=0.6.6",  # for converting QuerySets to PandasDataFrames
+    "dask-jobqueue>=0.7.3",  # for submitting on clusters
+    "dj-database-url>=0.5.0",  # for DigitalOcean URL conversion
+    "djangorestframework>=3.13.1",  # for our REST API
+    "django-filter>=21.1",  # sets up automatic filters for our REST API
+    "pyyaml>=6.0",  # for yaml configuration files
+    "plotly>=5.4.0",  # for interactive plots and visualization
+    "tqdm>=4.62.3",  # for monitoring progress of long for-loops
+    "pdoc>=11.0.0",  # for docs and markdown rendering in html templates
+    "nest_asyncio>=1.5.5",  # to allow async calls in Spyder/iPython/Jupyter
+    #
+    # For development and testing
[…]
